Created recipes.js in src/data folder, removed hard-coded recipe arrays from components and imported recipes.js instead
