+++ frontend/src/data/recipes.js
@@ -1,0 +1,4275 @@
+const recipes = {
+  topRated: [
+    {
+      id: 1,
+      title: "Delicious Pasta",
+      image: "../pasta.jpg",
+      rating: 4.8,
+      description:
+        "Creamy and flavorful pasta with a rich tomato sauce, garlic, and Parmesan cheese.",
+      prepTime: "15 minutes",
+      cookTime: "20 minutes",
+      servings: 4,
+      cuisine: "Italian",
+      category: "Main Course",
+      ingredients: [
+        { name: "Pasta (penne or spaghetti)", quantity: "12 ounces" },
+        { name: "Olive oil", quantity: "2 tablespoons" },
+        { name: "Garlic", quantity: "3 cloves", notes: "minced" },
+        { name: "Onion", quantity: "1 small", notes: "chopped" },
+        { name: "Tomato sauce", quantity: "2 cups" },
+        { name: "Heavy cream", quantity: "1/2 cup" },
+        { name: "Parmesan cheese", quantity: "1/2 cup", notes: "grated" },
+        { name: "Basil", quantity: "1/4 cup", notes: "chopped" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Black pepper", quantity: "to taste" },
+        {
+          name: "Red pepper flakes",
+          quantity: "1/2 teaspoon",
+          notes: "optional",
+        },
+      ],
+      instructions: [
+        "Bring a large pot of salted water to a boil and cook the pasta according to package instructions. Drain and set aside.",
+        "Heat olive oil in a pan over medium heat and sauté onions until soft. Add garlic and cook for 1 minute.",
+        "Pour in the tomato sauce and let it simmer for 5 minutes.",
+        "Stir in heavy cream, Parmesan cheese, salt, and pepper. Simmer for another 3-5 minutes until the sauce thickens.",
+        "Add the cooked pasta to the sauce and toss well to coat.",
+        "Garnish with fresh basil and additional Parmesan cheese.",
+        "Serve hot and enjoy!",
+      ],
+      nutritionalInfo: {
+        calories: "520",
+        fat: "22g",
+        protein: "18g",
+        carbs: "65g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Pasta",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 2,
+      title: "Spicy Tacos",
+      image: "../tacos.jpg",
+      rating: 4.7,
+      description:
+        "Flavorful and spicy tacos with seasoned ground beef, fresh toppings, and a zesty kick.",
+      prepTime: "20 minutes",
+      cookTime: "15 minutes",
+      servings: 4,
+      cuisine: "Mexican",
+      category: "Main Course",
+      ingredients: [
+        { name: "Ground beef", quantity: "1 pound" },
+        { name: "Taco seasoning", quantity: "2 tablespoons" },
+        { name: "Onion", quantity: "1 small", notes: "chopped" },
+        { name: "Garlic", quantity: "2 cloves", notes: "minced" },
+        { name: "Tomatoes", quantity: "1 cup", notes: "diced" },
+        { name: "Jalapeño", quantity: "1", notes: "sliced" },
+        { name: "Tortillas", quantity: "8 small" },
+        { name: "Cheddar cheese", quantity: "1 cup", notes: "shredded" },
+        { name: "Lettuce", quantity: "1 cup", notes: "shredded" },
+        { name: "Sour cream", quantity: "1/2 cup" },
+        { name: "Lime wedges", quantity: "for serving" },
+      ],
+      instructions: [
+        "Heat a skillet over medium heat and cook the chopped onion until soft.",
+        "Add the garlic and ground beef, cooking until the meat is browned.",
+        "Stir in taco seasoning and diced tomatoes. Simmer for 5 minutes.",
+        "Warm the tortillas in a dry pan or microwave.",
+        "Assemble tacos with seasoned beef, lettuce, cheese, jalapeños, and sour cream.",
+        "Serve with lime wedges on the side.",
+      ],
+      nutritionalInfo: {
+        calories: "450",
+        fat: "22g",
+        protein: "30g",
+        carbs: "40g",
+      },
+      dietaryInfo: ["Spicy", "Non-Vegetarian"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Beef",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 3,
+      title: "Pumpkin Cupcakes",
+      image: "../cakes.jpg",
+      rating: 4.8,
+      description:
+        "Soft and moist pumpkin cupcakes topped with a rich cream cheese frosting.",
+      prepTime: "15 minutes",
+      cookTime: "20 minutes",
+      servings: 12,
+      cuisine: "American",
+      category: "Dessert",
+      ingredients: [
+        { name: "All-purpose flour", quantity: "1 3/4 cups" },
+        { name: "Pumpkin puree", quantity: "1 cup" },
+        { name: "Granulated sugar", quantity: "3/4 cup" },
+        { name: "Brown sugar", quantity: "1/2 cup" },
+        { name: "Vegetable oil", quantity: "1/2 cup" },
+        { name: "Eggs", quantity: "2" },
+        { name: "Baking powder", quantity: "1 teaspoon" },
+        { name: "Baking soda", quantity: "1/2 teaspoon" },
+        { name: "Ground cinnamon", quantity: "1 teaspoon" },
+        { name: "Ground nutmeg", quantity: "1/2 teaspoon" },
+        { name: "Salt", quantity: "1/4 teaspoon" },
+        { name: "Vanilla extract", quantity: "1 teaspoon" },
+        {
+          name: "Cream cheese",
+          quantity: "8 ounces",
+          notes: "softened (for frosting)",
+        },
+        {
+          name: "Butter",
+          quantity: "1/2 cup",
+          notes: "softened (for frosting)",
+        },
+        { name: "Powdered sugar", quantity: "2 cups", notes: "(for frosting)" },
+      ],
+      instructions: [
+        "Preheat oven to 350°F (175°C) and line a cupcake tin with liners.",
+        "In a bowl, whisk together flour, baking powder, baking soda, cinnamon, nutmeg, and salt.",
+        "In another bowl, beat eggs, sugars, oil, pumpkin puree, and vanilla until smooth.",
+        "Gradually mix in dry ingredients until just combined.",
+        "Divide the batter evenly into cupcake liners, filling each about 2/3 full.",
+        "Bake for 18-20 minutes or until a toothpick inserted comes out clean.",
+        "Let cupcakes cool completely before frosting.",
+        "For frosting, beat cream cheese and butter until smooth, then mix in powdered sugar until fluffy.",
+        "Frost cupcakes and enjoy!",
+      ],
+      nutritionalInfo: {
+        calories: "320",
+        fat: "15g",
+        protein: "4g",
+        carbs: "42g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Dessert", "Snack"],
+      mainIngredient: "Pumpkin",
+      cookingMethod: "Baking",
+    },
+    {
+      id: 4,
+      title: "Apple Pie",
+      image: "../pie.jpg",
+      rating: 4.9,
+      description:
+        "A classic American apple pie with a flaky crust and a spiced apple filling.",
+      prepTime: "30 minutes",
+      cookTime: "50 minutes",
+      servings: 8,
+      cuisine: "American",
+      category: "Dessert",
+      ingredients: [
+        {
+          name: "Apples",
+          quantity: "6 cups",
+          notes: "peeled, cored, and sliced",
+        },
+        { name: "Granulated sugar", quantity: "3/4 cup" },
+        { name: "Brown sugar", quantity: "1/4 cup" },
+        { name: "All-purpose flour", quantity: "1/4 cup" },
+        { name: "Ground cinnamon", quantity: "1 teaspoon" },
+        { name: "Ground nutmeg", quantity: "1/4 teaspoon" },
+        { name: "Salt", quantity: "1/4 teaspoon" },
+        { name: "Lemon juice", quantity: "1 tablespoon" },
+        {
+          name: "Unsalted butter",
+          quantity: "2 tablespoons",
+          notes: "cut into small pieces",
+        },
+        { name: "Pie crust", quantity: "2", notes: "homemade or store-bought" },
+        { name: "Egg", quantity: "1", notes: "beaten (for egg wash)" },
+      ],
+      instructions: [
+        "Preheat oven to 375°F (190°C).",
+        "In a large bowl, mix apples with sugar, flour, cinnamon, nutmeg, salt, and lemon juice.",
+        "Roll out one pie crust and place it in a 9-inch pie dish.",
+        "Fill with apple mixture and dot with butter pieces.",
+        "Roll out the second crust and place it over the filling. Seal and crimp the edges.",
+        "Brush the top with egg wash and cut small slits to allow steam to escape.",
+        "Bake for 45-50 minutes, until the crust is golden brown and the filling is bubbling.",
+        "Allow the pie to cool for at least 2 hours before serving.",
+      ],
+      nutritionalInfo: {
+        calories: "350",
+        fat: "15g",
+        protein: "3g",
+        carbs: "50g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Dessert"],
+      mainIngredient: "Fruit",
+      cookingMethod: "Baking",
+    },
+    {
+      id: 5,
+      title: "Best Lasagna",
+      image: "../lasagna.jpg",
+      rating: 5.0,
+      description:
+        "A rich and cheesy Italian lasagna layered with a savory meat sauce and creamy ricotta filling.",
+      prepTime: "30 minutes",
+      cookTime: "1 hour",
+      servings: 8,
+      cuisine: "Italian",
+      category: "Main Course",
+      ingredients: [
+        { name: "Ground beef", quantity: "1 pound" },
+        {
+          name: "Italian sausage",
+          quantity: "1/2 pound",
+          notes: "mild or spicy",
+        },
+        { name: "Onion", quantity: "1 medium", notes: "chopped" },
+        { name: "Garlic", quantity: "3 cloves", notes: "minced" },
+        { name: "Tomato sauce", quantity: "28 ounces" },
+        { name: "Crushed tomatoes", quantity: "28 ounces" },
+        { name: "Tomato paste", quantity: "6 ounces" },
+        { name: "Water", quantity: "1/2 cup" },
+        { name: "Sugar", quantity: "1 tablespoon" },
+        { name: "Dried basil", quantity: "1 teaspoon" },
+        { name: "Dried oregano", quantity: "1 teaspoon" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Black pepper", quantity: "to taste" },
+        {
+          name: "Lasagna noodles",
+          quantity: "12",
+          notes: "cooked according to package instructions",
+        },
+        { name: "Ricotta cheese", quantity: "15 ounces" },
+        { name: "Egg", quantity: "1" },
+        { name: "Parmesan cheese", quantity: "1/2 cup", notes: "grated" },
+        { name: "Mozzarella cheese", quantity: "3 cups", notes: "shredded" },
+        { name: "Fresh parsley", quantity: "1/4 cup", notes: "chopped" },
+      ],
+      instructions: [
+        "In a large skillet, cook ground beef, sausage, and onion over medium heat until browned. Drain excess fat.",
+        "Add garlic and cook for 1 minute. Stir in tomato sauce, crushed tomatoes, tomato paste, water, sugar, basil, oregano, salt, and pepper. Simmer for 30 minutes.",
+        "In a bowl, mix ricotta cheese, egg, Parmesan cheese, and parsley.",
+        "Preheat oven to 375°F (190°C).",
+        "In a 9x13-inch baking dish, spread a layer of meat sauce. Add a layer of noodles, then spread a layer of ricotta mixture, followed by mozzarella cheese.",
+        "Repeat layers until all ingredients are used, finishing with mozzarella on top.",
+        "Cover with foil and bake for 25 minutes. Remove foil and bake for another 25 minutes until bubbly and golden brown.",
+        "Let rest for 10 minutes before slicing and serving.",
+      ],
+      nutritionalInfo: {
+        calories: "600",
+        fat: "35g",
+        protein: "40g",
+        carbs: "50g",
+      },
+      dietaryInfo: ["Non-Vegetarian"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Beef",
+      cookingMethod: "Baking",
+    },
+    {
+      id: 6,
+      title: "Harira",
+      image: "../harira.jpg",
+      rating: 4.8,
+      description:
+        "A traditional Moroccan soup made with tomatoes, lentils, chickpeas, and aromatic spices.",
+      prepTime: "15 minutes",
+      cookTime: "1 hour",
+      servings: 6,
+      cuisine: "Moroccan",
+      category: "Soup",
+      ingredients: [
+        { name: "Olive oil", quantity: "2 tablespoons" },
+        { name: "Onion", quantity: "1 large", notes: "chopped" },
+        { name: "Garlic", quantity: "3 cloves", notes: "minced" },
+        { name: "Celery", quantity: "2 stalks", notes: "chopped" },
+        { name: "Carrot", quantity: "1", notes: "chopped" },
+        { name: "Canned tomatoes", quantity: "28 ounces", notes: "crushed" },
+        { name: "Tomato paste", quantity: "2 tablespoons" },
+        { name: "Vegetable broth", quantity: "6 cups" },
+        { name: "Lentils", quantity: "1 cup", notes: "rinsed" },
+        {
+          name: "Chickpeas",
+          quantity: "1 can (15 ounces)",
+          notes: "drained and rinsed",
+        },
+        { name: "Ground cumin", quantity: "1 teaspoon" },
+        { name: "Ground coriander", quantity: "1 teaspoon" },
+        { name: "Ground cinnamon", quantity: "1/2 teaspoon" },
+        { name: "Ground turmeric", quantity: "1/2 teaspoon" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Black pepper", quantity: "to taste" },
+        { name: "Fresh cilantro", quantity: "1/4 cup", notes: "chopped" },
+        { name: "Fresh parsley", quantity: "1/4 cup", notes: "chopped" },
+        { name: "Lemon juice", quantity: "1 tablespoon" },
+      ],
+      instructions: [
+        "Heat olive oil in a large pot over medium heat. Add onions, garlic, celery, and carrot. Sauté until softened, about 5 minutes.",
+        "Stir in cumin, coriander, cinnamon, turmeric, salt, and pepper. Cook for 1 minute to release the flavors.",
+        "Add crushed tomatoes, tomato paste, and vegetable broth. Bring to a simmer.",
+        "Add lentils and chickpeas. Cover and cook for 45 minutes, stirring occasionally, until lentils are tender.",
+        "Stir in fresh cilantro, parsley, and lemon juice. Simmer for another 5 minutes.",
+        "Serve hot with crusty bread or over rice.",
+      ],
+      nutritionalInfo: {
+        calories: "300",
+        fat: "7g",
+        protein: "15g",
+        carbs: "45g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Stovetop",
+    },
+  ],
+  trending: [
+    {
+      id: 7,
+      title: "Vegan Curry",
+      image: "../curry.jpg",
+      rating: 4.9,
+      description:
+        "A hearty and flavorful vegan curry packed with vegetables, coconut milk, and aromatic spices.",
+      prepTime: "15 minutes",
+      cookTime: "40 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "trending",
+      ingredients: [
+        { name: "Olive oil", quantity: "2 tablespoons" },
+        { name: "Onion", quantity: "1 large", notes: "chopped" },
+        { name: "Garlic", quantity: "4 cloves", notes: "minced" },
+        { name: "Ginger", quantity: "1 tablespoon", notes: "grated" },
+        { name: "Carrot", quantity: "1", notes: "sliced" },
+        { name: "Bell pepper", quantity: "1", notes: "chopped" },
+        { name: "Sweet potato", quantity: "1 large", notes: "cubed" },
+        {
+          name: "Canned chickpeas",
+          quantity: "1 can (15 ounces)",
+          notes: "drained and rinsed",
+        },
+        { name: "Canned diced tomatoes", quantity: "1 can (14 ounces)" },
+        { name: "Coconut milk", quantity: "1 can (14 ounces)" },
+        { name: "Vegetable broth", quantity: "2 cups" },
+        { name: "Ground turmeric", quantity: "1 teaspoon" },
+        { name: "Ground cumin", quantity: "1 teaspoon" },
+        { name: "Ground coriander", quantity: "1 teaspoon" },
+        { name: "Garam masala", quantity: "1 teaspoon" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Black pepper", quantity: "to taste" },
+        { name: "Fresh spinach", quantity: "2 cups" },
+        { name: "Fresh cilantro", quantity: "1/4 cup", notes: "chopped" },
+        { name: "Lime juice", quantity: "1 tablespoon" },
+      ],
+      instructions: [
+        "Heat olive oil in a large pot over medium heat. Add onions and sauté until softened, about 5 minutes.",
+        "Add garlic, ginger, and spices (turmeric, cumin, coriander, garam masala). Cook for 1 minute to release flavors.",
+        "Stir in carrots, bell pepper, and sweet potato. Cook for another 3-4 minutes.",
+        "Add diced tomatoes, chickpeas, coconut milk, and vegetable broth. Bring to a boil, then reduce heat and simmer for 30 minutes until vegetables are tender.",
+        "Stir in fresh spinach and cook for 2 more minutes until wilted.",
+        "Remove from heat, stir in fresh cilantro and lime juice.",
+        "Serve hot with rice or naan.",
+      ],
+      nutritionalInfo: {
+        calories: "350",
+        fat: "15g",
+        protein: "10g",
+        carbs: "45g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "One-Pot Recipe",
+    },
+    {
+      id: 8,
+      title: "Chocolate Cake",
+      image: "../choco.jpg",
+      rating: 5.0,
+      description:
+        "A rich and moist chocolate cake with a creamy chocolate frosting, perfect for any occasion.",
+      prepTime: "20 minutes",
+      cookTime: "35 minutes",
+      servings: 8,
+      cuisine: "Dessert",
+      category: "trending",
+      ingredients: [
+        { name: "All-purpose flour", quantity: "1 and 3/4 cups" },
+        { name: "Granulated sugar", quantity: "2 cups" },
+        { name: "Unsweetened cocoa powder", quantity: "3/4 cup" },
+        { name: "Baking powder", quantity: "1 and 1/2 teaspoons" },
+        { name: "Baking soda", quantity: "1 and 1/2 teaspoons" },
+        { name: "Salt", quantity: "1 teaspoon" },
+        { name: "Eggs", quantity: "2 large" },
+        { name: "Whole milk", quantity: "1 cup" },
+        { name: "Vegetable oil", quantity: "1/2 cup" },
+        { name: "Vanilla extract", quantity: "2 teaspoons" },
+        { name: "Boiling water", quantity: "1 cup" },
+      ],
+      instructions: [
+        "Preheat oven to 350°F (175°C). Grease and flour two 9-inch round cake pans.",
+        "In a large bowl, whisk together flour, sugar, cocoa powder, baking powder, baking soda, and salt.",
+        "Add eggs, milk, vegetable oil, and vanilla extract. Beat on medium speed until well combined.",
+        "Reduce mixer speed to low and carefully add boiling water. Mix until the batter is smooth.",
+        "Divide batter evenly between the prepared cake pans.",
+        "Bake for 30-35 minutes or until a toothpick inserted in the center comes out clean.",
+        "Cool cakes in pans for 10 minutes, then transfer to a wire rack to cool completely before frosting.",
+        "Frost with your favorite chocolate frosting and enjoy!",
+      ],
+      nutritionalInfo: {
+        calories: "450",
+        fat: "18g",
+        protein: "6g",
+        carbs: "65g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Dessert"],
+      mainIngredient: "Chocolate",
+      cookingMethod: "Baking",
+    },
+    {
+      id: 9,
+      title: "Corn Fritters",
+      image: "../corn.jpg",
+      rating: 4.7,
+      description:
+        "Crispy and golden corn fritters made with fresh corn kernels, perfect as a snack or appetizer.",
+      prepTime: "15 minutes",
+      cookTime: "10 minutes",
+      servings: 4,
+      cuisine: "American",
+      category: "trending",
+      ingredients: [
+        { name: "Fresh corn kernels", quantity: "1 cup" },
+        { name: "All-purpose flour", quantity: "1/2 cup" },
+        { name: "Egg", quantity: "1 large" },
+        { name: "Milk", quantity: "1/4 cup" },
+        { name: "Baking powder", quantity: "1 teaspoon" },
+        { name: "Salt", quantity: "1/2 teaspoon" },
+        { name: "Black pepper", quantity: "1/4 teaspoon" },
+        { name: "Chopped green onions", quantity: "1/4 cup" },
+        { name: "Shredded cheddar cheese", quantity: "1/2 cup" },
+        { name: "Vegetable oil", quantity: "for frying" },
+      ],
+      instructions: [
+        "In a bowl, whisk together flour, baking powder, salt, and black pepper.",
+        "In another bowl, beat the egg and mix in the milk.",
+        "Combine the wet and dry ingredients, then fold in corn kernels, green onions, and shredded cheese.",
+        "Heat vegetable oil in a pan over medium heat.",
+        "Drop spoonfuls of batter into the pan and flatten slightly.",
+        "Fry for 2-3 minutes on each side until golden brown and crispy.",
+        "Drain on paper towels and serve warm with your favorite dipping sauce.",
+      ],
+      nutritionalInfo: {
+        calories: "220",
+        fat: "12g",
+        protein: "5g",
+        carbs: "22g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Snacks", "Appetizers"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 10,
+      title: "Bread Cheese Lollipop",
+      image: "../lollipop.jpg",
+      rating: 4.6,
+      description:
+        "Crispy, cheesy bread lollipops with a golden crust, perfect for a fun snack or party appetizer.",
+      prepTime: "20 minutes",
+      cookTime: "10 minutes",
+      servings: 6,
+      cuisine: "Fusion",
+      category: "trending",
+      ingredients: [
+        { name: "Bread slices", quantity: "4 slices" },
+        { name: "Mozzarella cheese", quantity: "1 cup", notes: "grated" },
+        { name: "Boiled potato", quantity: "1 medium", notes: "mashed" },
+        { name: "Cornflour", quantity: "2 tablespoons" },
+        { name: "Black pepper", quantity: "1/2 teaspoon" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Chili flakes", quantity: "1/2 teaspoon" },
+        { name: "Oregano", quantity: "1/2 teaspoon" },
+        { name: "Bread crumbs", quantity: "1/2 cup" },
+        { name: "Milk", quantity: "1/4 cup" },
+        { name: "Oil", quantity: "for deep frying" },
+        { name: "Wooden skewers", quantity: "6 pieces" },
+      ],
+      instructions: [
+        "Cut the bread slices into small pieces and soak them in milk for a few seconds.",
+        "Squeeze out excess milk and mix the bread with mashed potato, cheese, cornflour, salt, pepper, chili flakes, and oregano.",
+        "Shape the mixture into small balls and insert a wooden skewer into each ball.",
+        "Roll the lollipops in bread crumbs for a crispy coating.",
+        "Heat oil in a pan and deep fry until golden brown and crispy.",
+        "Drain on paper towels and serve hot with ketchup or a dipping sauce of your choice.",
+      ],
+      nutritionalInfo: {
+        calories: "250",
+        fat: "14g",
+        protein: "7g",
+        carbs: "24g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Snacks", "Appetizers"],
+      mainIngredient: "Cheese",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 11,
+      title: "Sweet Potato Boats",
+      image: "../boats.jpg",
+      rating: 4.7,
+      description:
+        "Baked sweet potatoes stuffed with a delicious mix of veggies, cheese, and spices for a healthy and flavorful meal.",
+      prepTime: "15 minutes",
+      cookTime: "45 minutes",
+      servings: 4,
+      cuisine: "American",
+      category: "trending",
+      ingredients: [
+        { name: "Sweet potatoes", quantity: "4 medium" },
+        { name: "Olive oil", quantity: "2 tablespoons" },
+        { name: "Garlic", quantity: "2 cloves", notes: "minced" },
+        { name: "Red bell pepper", quantity: "1 small", notes: "chopped" },
+        { name: "Black beans", quantity: "1 cup", notes: "cooked" },
+        { name: "Corn", quantity: "1/2 cup" },
+        { name: "Cherry tomatoes", quantity: "1/2 cup", notes: "halved" },
+        { name: "Spinach", quantity: "1/2 cup", notes: "chopped" },
+        { name: "Cumin", quantity: "1 teaspoon" },
+        { name: "Paprika", quantity: "1 teaspoon" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Black pepper", quantity: "to taste" },
+        { name: "Cheddar cheese", quantity: "1/2 cup", notes: "shredded" },
+        { name: "Sour cream", quantity: "for topping" },
+        { name: "Cilantro", quantity: "for garnish" },
+      ],
+      instructions: [
+        "Preheat oven to 400°F (200°C) and line a baking sheet with parchment paper.",
+        "Wash the sweet potatoes, poke holes with a fork, and rub with olive oil.",
+        "Bake for 40-45 minutes or until tender.",
+        "While sweet potatoes are baking, heat olive oil in a pan over medium heat.",
+        "Sauté garlic and red bell pepper for 2-3 minutes.",
+        "Add black beans, corn, tomatoes, spinach, cumin, paprika, salt, and pepper. Cook for another 5 minutes.",
+        "Once the sweet potatoes are cooked, let them cool slightly, then slice them in half lengthwise.",
+        "Scoop out some of the flesh, mash it, and mix it with the sautéed veggie mixture.",
+        "Stuff the mixture back into the sweet potato skins, top with shredded cheese, and bake for another 5 minutes until the cheese melts.",
+        "Garnish with sour cream and cilantro before serving.",
+      ],
+      nutritionalInfo: {
+        calories: "320",
+        fat: "10g",
+        protein: "9g",
+        carbs: "50g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Potato",
+      cookingMethod: "Baking",
+    },
+    {
+      id: 12,
+      title: "Walnut Chikki",
+      image: "../chikki.jpg",
+      rating: 4.8,
+      description:
+        "A crunchy and nutty Indian sweet made with walnuts and jaggery, perfect for a healthy energy boost.",
+      prepTime: "10 minutes",
+      cookTime: "15 minutes",
+      servings: 10,
+      cuisine: "Indian",
+      category: "trending",
+      ingredients: [
+        { name: "Walnuts", quantity: "1 cup", notes: "chopped" },
+        { name: "Jaggery", quantity: "3/4 cup", notes: "grated" },
+        { name: "Ghee", quantity: "1 teaspoon" },
+        { name: "Cardamom powder", quantity: "1/2 teaspoon" },
+      ],
+      instructions: [
+        "Grease a plate or parchment paper with ghee and set aside.",
+        "Heat a pan over medium heat and dry roast the walnuts for 2-3 minutes. Set aside.",
+        "In the same pan, add ghee and jaggery. Stir continuously until the jaggery melts and forms a syrup.",
+        "To check if the syrup is ready, drop a small amount into cold water; it should form a hard ball.",
+        "Add roasted walnuts and cardamom powder to the syrup. Mix quickly and pour onto the greased plate.",
+        "Flatten with a rolling pin or spatula and let it cool for 10 minutes.",
+        "Once firm, break into pieces and store in an airtight container.",
+      ],
+      nutritionalInfo: {
+        calories: "150",
+        fat: "9g",
+        protein: "3g",
+        carbs: "18g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Dessert", "Snacks"],
+      mainIngredient: "Nuts",
+      cookingMethod: "Stovetop",
+    },
+  ],
+  newest: [
+    {
+      id: 13,
+      title: "Summer Salad",
+      image: "../salad.jpg",
+      rating: 4.7,
+      description:
+        "A refreshing and vibrant salad packed with fresh fruits, greens, and a zesty dressing, perfect for hot summer days.",
+      prepTime: "15 minutes",
+      cookTime: "0 minutes",
+      servings: 4,
+      cuisine: "International",
+      category: "Salad",
+      ingredients: [
+        { name: "Mixed greens", quantity: "4 cups" },
+        { name: "Strawberries", quantity: "1 cup", notes: "sliced" },
+        { name: "Blueberries", quantity: "1/2 cup" },
+        { name: "Mango", quantity: "1 cup", notes: "diced" },
+        { name: "Feta cheese", quantity: "1/2 cup", notes: "crumbled" },
+        { name: "Walnuts", quantity: "1/4 cup", notes: "toasted" },
+        { name: "Balsamic vinegar", quantity: "2 tablespoons" },
+        { name: "Olive oil", quantity: "3 tablespoons" },
+        { name: "Honey", quantity: "1 tablespoon" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Black pepper", quantity: "to taste" },
+      ],
+      instructions: [
+        "In a large bowl, combine mixed greens, strawberries, blueberries, mango, feta cheese, and toasted walnuts.",
+        "In a small bowl, whisk together balsamic vinegar, olive oil, honey, salt, and black pepper to make the dressing.",
+        "Drizzle the dressing over the salad and toss gently to combine.",
+        "Serve immediately and enjoy a refreshing summer delight!",
+      ],
+      nutritionalInfo: {
+        calories: "180",
+        fat: "10g",
+        protein: "5g",
+        carbs: "20g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner", "Snacks"],
+      mainIngredient: "Fruit",
+      cookingMethod: "No-Cook",
+    },
+    {
+      id: 14,
+      title: "Grilled Salmon",
+      image: "../salmon.jpg",
+      rating: 4.9,
+      description:
+        "A delicious and healthy grilled salmon dish with a tangy lemon and herb marinade.",
+      prepTime: "10 minutes",
+      cookTime: "15 minutes",
+      servings: 2,
+      cuisine: "American",
+      category: "Main Course",
+      ingredients: [
+        { name: "Salmon fillets", quantity: "2", notes: "6 oz each" },
+        { name: "Olive oil", quantity: "2 tablespoons" },
+        { name: "Lemon juice", quantity: "2 tablespoons" },
+        { name: "Garlic", quantity: "2 cloves", notes: "minced" },
+        { name: "Fresh dill", quantity: "1 tablespoon", notes: "chopped" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Black pepper", quantity: "to taste" },
+        { name: "Paprika", quantity: "1 teaspoon" },
+        { name: "Honey", quantity: "1 teaspoon" },
+      ],
+      instructions: [
+        "In a small bowl, whisk together olive oil, lemon juice, garlic, dill, salt, black pepper, paprika, and honey to make the marinade.",
+        "Place salmon fillets in a shallow dish and pour the marinade over them. Let it sit for at least 10 minutes.",
+        "Preheat the grill to medium-high heat and lightly oil the grates.",
+        "Grill the salmon for 5-7 minutes per side, or until the fish flakes easily with a fork.",
+        "Remove from the grill and serve immediately with lemon wedges.",
+      ],
+      nutritionalInfo: {
+        calories: "320",
+        fat: "18g",
+        protein: "34g",
+        carbs: "5g",
+      },
+      dietaryInfo: ["High-Protein", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Fish",
+      cookingMethod: "Grilling",
+    },
+    {
+      id: 15,
+      title: "Loco Moco",
+      image: "../loco.jpg",
+      rating: 4.8,
+      description:
+        "A Hawaiian comfort food classic featuring a juicy hamburger patty, rich brown gravy, and a fried egg over rice.",
+      prepTime: "10 minutes",
+      cookTime: "20 minutes",
+      servings: 2,
+      cuisine: "Hawaiian",
+      category: "Main Course",
+      ingredients: [
+        { name: "Ground beef", quantity: "½ pound", notes: "80/20 lean" },
+        { name: "Salt", quantity: "½ teaspoon" },
+        { name: "Black pepper", quantity: "½ teaspoon" },
+        { name: "Worcestershire sauce", quantity: "1 teaspoon" },
+        { name: "Eggs", quantity: "2" },
+        { name: "Cooked white rice", quantity: "2 cups" },
+        { name: "Butter", quantity: "1 tablespoon" },
+        { name: "Onion", quantity: "½ cup", notes: "finely chopped" },
+        { name: "Beef broth", quantity: "1 cup" },
+        { name: "Soy sauce", quantity: "1 tablespoon" },
+        {
+          name: "Cornstarch",
+          quantity: "1 teaspoon",
+          notes: "mixed with 2 tbsp water",
+        },
+        {
+          name: "Green onions",
+          quantity: "2 tablespoons",
+          notes: "chopped, for garnish",
+        },
+      ],
+      instructions: [
+        "In a bowl, mix ground beef with salt, black pepper, and Worcestershire sauce. Form into two patties.",
+        "Heat a skillet over medium-high heat and cook the patties for about 4-5 minutes per side, until browned and cooked through.",
+        "Remove patties and set aside. In the same skillet, melt butter and sauté onions until softened.",
+        "Add beef broth and soy sauce, then stir in cornstarch mixture to thicken the gravy. Simmer for 2 minutes.",
+        "In a separate pan, fry the eggs sunny-side up.",
+        "To serve, place a scoop of rice on each plate, top with a beef patty, spoon over the gravy, and finish with a fried egg.",
+        "Garnish with chopped green onions and serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: "550",
+        fat: "28g",
+        protein: "35g",
+        carbs: "45g",
+      },
+      dietaryInfo: ["High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Beef",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 16,
+      title: "Cinnamon Roll French Toast Casserole",
+      image: "../toast.jpg",
+      rating: 4.9,
+      description:
+        "A deliciously sweet and comforting breakfast casserole made with cinnamon rolls, eggs, and a rich glaze.",
+      prepTime: "10 minutes",
+      cookTime: "35 minutes",
+      servings: 6,
+      cuisine: "American",
+      category: "Breakfast",
+      ingredients: [
+        {
+          name: "Refrigerated cinnamon rolls",
+          quantity: "2 cans",
+          notes: "with icing",
+        },
+        { name: "Eggs", quantity: "4" },
+        { name: "Milk", quantity: "½ cup" },
+        { name: "Vanilla extract", quantity: "1 teaspoon" },
+        { name: "Ground cinnamon", quantity: "1 teaspoon" },
+        { name: "Maple syrup", quantity: "¼ cup" },
+        { name: "Butter", quantity: "2 tablespoons", notes: "melted" },
+        {
+          name: "Powdered sugar",
+          quantity: "2 tablespoons",
+          notes: "for dusting (optional)",
+        },
+      ],
+      instructions: [
+        "Preheat oven to 350°F (175°C) and grease a baking dish.",
+        "Cut the cinnamon rolls into small pieces and spread them evenly in the baking dish.",
+        "In a bowl, whisk together eggs, milk, vanilla extract, ground cinnamon, and maple syrup.",
+        "Pour the egg mixture over the cinnamon rolls and gently stir to coat.",
+        "Drizzle melted butter over the top and bake for 30-35 minutes, or until golden brown.",
+        "Let cool slightly, then drizzle with the icing from the cinnamon roll package.",
+        "Dust with powdered sugar if desired and serve warm.",
+      ],
+      nutritionalInfo: {
+        calories: "480",
+        fat: "20g",
+        protein: "7g",
+        carbs: "68g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Breakfast"],
+      mainIngredient: "Milk",
+      cookingMethod: "Baking",
+    },
+    {
+      id: 17,
+      title: "Frikadellen (German Meat Patties)",
+      image: "../Frikadellen.jpg",
+      rating: 4.7,
+      description:
+        "A traditional German dish made with seasoned ground meat, pan-fried to perfection.",
+      prepTime: "15 minutes",
+      cookTime: "20 minutes",
+      servings: 4,
+      cuisine: "German",
+      category: "Main Course",
+      ingredients: [
+        { name: "Ground beef", quantity: "1 pound" },
+        { name: "Ground pork", quantity: "½ pound" },
+        {
+          name: "Bread roll",
+          quantity: "1",
+          notes: "soaked in water and squeezed dry",
+        },
+        { name: "Egg", quantity: "1" },
+        { name: "Onion", quantity: "1 small", notes: "finely chopped" },
+        { name: "Garlic", quantity: "2 cloves", notes: "minced" },
+        { name: "Mustard", quantity: "1 tablespoon" },
+        { name: "Salt", quantity: "1 teaspoon" },
+        { name: "Black pepper", quantity: "½ teaspoon" },
+        { name: "Paprika", quantity: "½ teaspoon" },
+        { name: "Parsley", quantity: "2 tablespoons", notes: "chopped" },
+        {
+          name: "Vegetable oil",
+          quantity: "2 tablespoons",
+          notes: "for frying",
+        },
+      ],
+      instructions: [
+        "In a large bowl, combine ground beef, ground pork, soaked bread roll, egg, onion, garlic, mustard, salt, pepper, paprika, and parsley.",
+        "Mix well with your hands until fully combined.",
+        "Shape the mixture into 4-6 patties, flattening slightly.",
+        "Heat vegetable oil in a large skillet over medium heat.",
+        "Fry the patties for about 5 minutes per side, or until golden brown and cooked through.",
+        "Serve hot with mustard, bread, or potato salad.",
+      ],
+      nutritionalInfo: {
+        calories: "350",
+        fat: "22g",
+        protein: "30g",
+        carbs: "10g",
+      },
+      dietaryInfo: ["High Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Beef",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 18,
+      title: "Coffee Jelly",
+      image: "../jelly.jpg",
+      rating: 4.8,
+      description:
+        "A refreshing Japanese dessert made with coffee-flavored gelatin cubes served with sweet cream.",
+      prepTime: "10 minutes",
+      cookTime: "5 minutes",
+      servings: 4,
+      cuisine: "Japanese",
+      category: "Dessert",
+      ingredients: [
+        { name: "Strong brewed coffee", quantity: "2 cups" },
+        { name: "Gelatin powder", quantity: "2 teaspoons" },
+        { name: "Sugar", quantity: "2 tablespoons" },
+        { name: "Water", quantity: "½ cup" },
+        {
+          name: "Sweetened condensed milk",
+          quantity: "½ cup",
+          notes: "or heavy cream for serving",
+        },
+      ],
+      instructions: [
+        "In a small bowl, sprinkle gelatin over ¼ cup of water and let it bloom for 5 minutes.",
+        "In a saucepan, heat brewed coffee and sugar over medium heat until sugar dissolves.",
+        "Add the bloomed gelatin mixture and stir until fully dissolved.",
+        "Remove from heat and pour into a shallow dish or mold.",
+        "Refrigerate for at least 2 hours or until set.",
+        "Cut into small cubes and serve in a glass with sweetened condensed milk or cream.",
+      ],
+      nutritionalInfo: {
+        calories: "120",
+        fat: "4g",
+        protein: "3g",
+        carbs: "20g",
+      },
+      dietaryInfo: ["Vegetarian", "Low Fat"],
+      mealType: ["Dessert", "Snacks"],
+      mainIngredient: "Milk",
+      cookingMethod: "No-Cook",
+    },
+  ],
+
+  breakfast: [
+    {
+      id: 19,
+      title: "Poha",
+      image: "https://www.spiceupthecurry.com/wp-content/uploads/2014/04/batata-poha-recipe-1-500x500.jpg",
+      rating: 4.7,
+      description:
+        "A light and flavorful Indian breakfast dish made with flattened rice, spices, and peanuts.",
+      prepTime: "10 minutes",
+      cookTime: "10 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Breakfast",
+      ingredients: [
+        { name: "Flattened rice (poha)", quantity: "1 cup" },
+        { name: "Oil", quantity: "2 tablespoons" },
+        { name: "Mustard seeds", quantity: "1 teaspoon" },
+        { name: "Cumin seeds", quantity: "1/2 teaspoon" },
+        { name: "Green chili", quantity: "1", notes: "chopped" },
+        { name: "Onion", quantity: "1 small", notes: "finely chopped" },
+        { name: "Turmeric powder", quantity: "1/2 teaspoon" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Sugar", quantity: "1 teaspoon" },
+        { name: "Lemon juice", quantity: "1 tablespoon" },
+        { name: "Peanuts", quantity: "2 tablespoons", notes: "roasted" },
+        {
+          name: "Coriander leaves",
+          quantity: "2 tablespoons",
+          notes: "chopped",
+        },
+        { name: "Curry leaves", quantity: "6-7" },
+      ],
+      instructions: [
+        "Rinse poha in water and drain excess water. Set aside for 5 minutes.",
+        "Heat oil in a pan, add mustard seeds and cumin seeds. Let them splutter.",
+        "Add chopped green chili, curry leaves, and onions. Sauté until onions turn translucent.",
+        "Add turmeric powder, salt, and sugar. Stir well.",
+        "Add the soaked poha and mix gently to coat with spices.",
+        "Sprinkle lemon juice and mix again.",
+        "Garnish with roasted peanuts and fresh coriander leaves.",
+        "Serve hot with tea or yogurt.",
+      ],
+      nutritionalInfo: {
+        calories: "250",
+        fat: "8g",
+        protein: "6g",
+        carbs: "40g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Breakfast"],
+      mainIngredient: "Rice",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 20,
+      title: "Aloo Paratha",
+      image:
+        "https://www.indianhealthyrecipes.com/wp-content/uploads/2020/08/aloo-paratha-recipe.jpg",
+      rating: 4.7,
+      description:
+        "A stuffed Indian flatbread made with spiced mashed potatoes and whole wheat flour, served hot with butter, curd, or pickle.",
+      prepTime: "20 minutes",
+      cookTime: "15 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Breakfast",
+      ingredients: [
+        { name: "Whole wheat flour", quantity: "2 cups" },
+        { name: "Boiled potatoes", quantity: "3 medium", notes: "mashed" },
+        { name: "Green chili", quantity: "1-2", notes: "finely chopped" },
+        { name: "Coriander leaves", quantity: "2 tbsp", notes: "chopped" },
+        { name: "Red chili powder", quantity: "1/2 tsp" },
+        { name: "Garam masala", quantity: "1/2 tsp" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Oil/Ghee", quantity: "as needed", notes: "for roasting" },
+      ],
+      instructions: [
+        "Prepare the dough with whole wheat flour and water. Rest for 15 minutes.",
+        "Mix mashed potatoes with green chili, coriander, red chili powder, garam masala, and salt.",
+        "Roll out dough, stuff with potato mixture, and seal the edges.",
+        "Roll again gently into a flatbread.",
+        "Cook on a hot pan with oil or ghee until golden brown on both sides.",
+        "Serve hot with curd, butter, or pickle.",
+      ],
+      nutritionalInfo: {
+        calories: "320",
+        fat: "12g",
+        protein: "7g",
+        carbs: "45g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Breakfast"],
+      mainIngredient: "Potato",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 21,
+      title: "Upma",
+      image: "https://yennadosa.com/wp-content/uploads/2024/07/yenna-blog.png",
+      rating: 4.5,
+      description:
+        "A savory South Indian breakfast made with semolina (rava), vegetables, and mild spices.",
+      prepTime: "10 minutes",
+      cookTime: "15 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Breakfast",
+      ingredients: [
+        { name: "Semolina (rava/sooji)", quantity: "1 cup" },
+        { name: "Oil or ghee", quantity: "2 tbsp" },
+        { name: "Mustard seeds", quantity: "1/2 tsp" },
+        { name: "Green chili", quantity: "1", notes: "chopped" },
+        { name: "Curry leaves", quantity: "6-7" },
+        { name: "Onion", quantity: "1 small", notes: "finely chopped" },
+        {
+          name: "Mixed vegetables",
+          quantity: "1/2 cup",
+          notes: "carrot, beans, peas",
+        },
+        { name: "Water", quantity: "2.5 cups" },
+        { name: "Salt", quantity: "to taste" },
+      ],
+      instructions: [
+        "Dry roast semolina until light golden and aromatic. Set aside.",
+        "Heat oil, add mustard seeds, curry leaves, and green chili.",
+        "Add onion and sauté until translucent. Add vegetables and cook briefly.",
+        "Add water and salt. Bring to a boil.",
+        "Slowly add semolina while stirring continuously to avoid lumps.",
+        "Cook until the mixture thickens. Serve hot with chutney.",
+      ],
+      nutritionalInfo: {
+        calories: "250",
+        fat: "8g",
+        protein: "6g",
+        carbs: "40g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Breakfast"],
+      mainIngredient: "Semolina",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 22,
+      title: "Dosa",
+      image:
+        "https://delishglobe.com/wp-content/uploads/2024/09/Masala-dosa-1.png",
+      rating: 4.8,
+      description:
+        "A crispy and thin South Indian crepe made from fermented rice and lentil batter, often served with chutney and sambar.",
+      prepTime: "15 minutes (excluding fermentation)",
+      cookTime: "10 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Breakfast",
+      ingredients: [
+        { name: "Rice", quantity: "2 cups" },
+        { name: "Urad dal (split black gram)", quantity: "1/2 cup" },
+        { name: "Fenugreek seeds", quantity: "1 tsp" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Oil", quantity: "as needed", notes: "for roasting" },
+      ],
+      instructions: [
+        "Soak rice, urad dal, and fenugreek seeds for 6 hours.",
+        "Grind into a smooth batter and ferment overnight.",
+        "Add salt and mix well.",
+        "Pour a ladle of batter onto a hot pan, spread thinly.",
+        "Drizzle oil around edges and cook until golden brown.",
+        "Serve hot with chutney and sambar.",
+      ],
+      nutritionalInfo: {
+        calories: "200",
+        fat: "5g",
+        protein: "6g",
+        carbs: "36g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Breakfast"],
+      mainIngredient: "Rice",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 23,
+      title: "Idli",
+      image:
+        "https://vaya.in/recipes/wp-content/uploads/2018/02/Idli-and-Sambar-1.jpg",
+      rating: 4.6,
+      description:
+        "Soft and fluffy steamed rice cakes from South India, typically served with chutney and sambar.",
+      prepTime: "15 minutes (excluding fermentation)",
+      cookTime: "10 minutes",
+      servings: 3,
+      cuisine: "Indian",
+      category: "Breakfast",
+      ingredients: [
+        { name: "Rice", quantity: "2 cups" },
+        { name: "Urad dal (split black gram)", quantity: "1/2 cup" },
+        { name: "Fenugreek seeds", quantity: "1 tsp" },
+        { name: "Salt", quantity: "to taste" },
+      ],
+      instructions: [
+        "Soak rice, urad dal, and fenugreek seeds for 6 hours.",
+        "Grind separately, then mix and ferment overnight.",
+        "Add salt and stir well.",
+        "Pour batter into idli molds and steam for 10 minutes.",
+        "Serve hot with chutney and sambar.",
+      ],
+      nutritionalInfo: {
+        calories: "180",
+        fat: "1g",
+        protein: "5g",
+        carbs: "38g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Breakfast"],
+      mainIngredient: "Rice",
+      cookingMethod: "Steaming",
+    },
+    {
+      id: 24,
+      title: "Besan Chilla",
+      image:
+        "https://i.ytimg.com/vi/KeEMDKsI0rU/hq720.jpg?sqp=-oaymwEhCK4FEIIDSFryq4qpAxMIARUAAAAAGAElAADIQj0AgKJD&rs=AOn4CLDurg-qEHjB1whS_78b6NmnCmefPw",
+      rating: 4.4,
+      description:
+        "A savory pancake made from gram flour (besan), spices, and herbs, served as a quick breakfast or snack.",
+      prepTime: "10 minutes",
+      cookTime: "10 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Breakfast",
+      ingredients: [
+        { name: "Gram flour (besan)", quantity: "1 cup" },
+        { name: "Onion", quantity: "1 small", notes: "finely chopped" },
+        { name: "Tomato", quantity: "1 small", notes: "finely chopped" },
+        { name: "Green chili", quantity: "1", notes: "chopped" },
+        { name: "Coriander leaves", quantity: "2 tbsp", notes: "chopped" },
+        { name: "Turmeric powder", quantity: "1/4 tsp" },
+        { name: "Red chili powder", quantity: "1/2 tsp" },
+        { name: "Salt", quantity: "to taste" },
+        { name: "Water", quantity: "as required", notes: "for batter" },
+        { name: "Oil", quantity: "as needed", notes: "for cooking" },
+      ],
+      instructions: [
+        "Mix gram flour, water, salt, turmeric, and red chili powder to make a smooth batter.",
+        "Add onion, tomato, green chili, and coriander leaves.",
+        "Heat oil on a pan, pour batter, and spread into a pancake.",
+        "Cook until golden brown on both sides.",
+        "Serve hot with chutney or curd.",
+      ],
+      nutritionalInfo: {
+        calories: "220",
+        fat: "7g",
+        protein: "8g",
+        carbs: "30g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Breakfast", "Snacks"],
+      mainIngredient: "Gram Flour",
+      cookingMethod: "Stovetop",
+    },
+  ],
+  lunch: [
+    {
+      id: 25,
+      title: "Butter Chicken",
+      image:
+        "https://www.spiceroots.com/spiceroots/wp-content/uploads/2008/05/butterchicken-1024x682-1.jpg",
+      rating: 4.9,
+      prepTime: "40 minutes",
+      cookTime: "30 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Rich and creamy tomato-based chicken curry",
+      ingredients: [
+        { quantity: "500g", name: "chicken", notes: "boneless, diced" },
+        { quantity: "2 tablespoons", name: "butter" },
+        { quantity: "1 cup", name: "tomato puree" },
+        { quantity: "1/2 cup", name: "cream" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "ginger garlic paste" },
+        { quantity: "1 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Marinate chicken with ginger garlic paste, salt, and garam masala for 20 minutes.",
+        "Heat butter in a pan and sauté the chicken until partially cooked.",
+        "Add tomato puree and cook until the oil separates.",
+        "Stir in cream and simmer until chicken is fully cooked and gravy is creamy.",
+        "Garnish with coriander leaves and serve with naan or rice.",
+      ],
+      nutritionalInfo: {
+        calories: "420",
+        fat: "28g",
+        protein: "28g",
+        carbs: "10g",
+      },
+      dietaryInfo: ["Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 26,
+      title: "Paneer Tikka Masala",
+      image:
+        "https://www.nehascookbook.com/wp-content/uploads/2022/11/Paneer-tikka-masala-WS-500x500.jpg",
+      rating: 4.8,
+      prepTime: "35 minutes",
+      cookTime: "25 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Grilled paneer cubes cooked in a spiced tomato gravy",
+      ingredients: [
+        { quantity: "250g", name: "paneer", notes: "cubed" },
+        { quantity: "1 cup", name: "yogurt" },
+        { quantity: "1 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "turmeric powder" },
+        { quantity: "1/2 cup", name: "tomato puree" },
+        { quantity: "1 teaspoon", name: "cumin powder" },
+        { quantity: "1/2 cup", name: "cream" },
+      ],
+      instructions: [
+        "Marinate paneer with yogurt, chili powder, turmeric, and cumin powder. Rest for 20 minutes.",
+        "Grill or pan-fry the paneer until golden.",
+        "Heat oil in a pan, add tomato puree, and cook until thickened.",
+        "Add cream and the grilled paneer, simmer for a few minutes.",
+        "Serve hot with naan, roti, or rice.",
+      ],
+      nutritionalInfo: {
+        calories: "380",
+        fat: "24g",
+        protein: "15g",
+        carbs: "18g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Paneer",
+      cookingMethod: "Grilling",
+    },
+    {
+      id: 27,
+      title: "Vegetable Biryani",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRPbQMgD79WJwULTkrE5C-MsHcUMyKozMq-Kw&s",
+      rating: 4.7,
+      prepTime: "45 minutes",
+      cookTime: "30 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Fragrant rice cooked with mixed vegetables and spices",
+      ingredients: [
+        { quantity: "2 cups", name: "basmati rice" },
+        { quantity: "1 cup", name: "mixed vegetables", notes: "chopped" },
+        { quantity: "1", name: "onion", notes: "sliced" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "cumin seeds" },
+        { quantity: "1/2 teaspoon", name: "turmeric powder" },
+        { quantity: "1 cup", name: "water" },
+      ],
+      instructions: [
+        "Soak rice for 30 minutes and drain.",
+        "Heat oil, add cumin seeds and sliced onion, sauté until golden.",
+        "Add vegetables, turmeric, and garam masala, cook briefly.",
+        "Add rice, water, and salt. Cook on low heat until rice is fluffy.",
+        "Serve hot with raita or curry.",
+      ],
+      nutritionalInfo: {
+        calories: "350",
+        fat: "10g",
+        protein: "8g",
+        carbs: "58g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "One-Pot Recipe",
+    },
+    {
+      id: 28,
+      title: "Dal Tadka",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR4kLMn_YJz4DEPNat_gUTF_QWOxqKiIRWMQw&s",
+      rating: 4.8,
+      prepTime: "30 minutes",
+      cookTime: "25 minutes",
+      servings: 3,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Yellow lentils tempered with garlic and spices",
+      ingredients: [
+        { quantity: "1 cup", name: "toor dal (pigeon peas)" },
+        { quantity: "2 cups", name: "water" },
+        { quantity: "1 teaspoon", name: "turmeric powder" },
+        { quantity: "1 teaspoon", name: "mustard seeds" },
+        { quantity: "2 cloves", name: "garlic", notes: "minced" },
+        { quantity: "1 teaspoon", name: "ghee" },
+        { quantity: "1", name: "green chili", notes: "chopped" },
+      ],
+      instructions: [
+        "Pressure cook dal with turmeric and water until soft.",
+        "Mash slightly and simmer with salt.",
+        "Heat ghee, add mustard seeds, garlic, and green chili.",
+        "Pour tempering over dal and stir well.",
+        "Serve with steamed rice or roti.",
+      ],
+      nutritionalInfo: {
+        calories: "210",
+        fat: "6g",
+        protein: "10g",
+        carbs: "30g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free", "High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Dal",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 29,
+      title: "Chole (Chickpea Curry)",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSnoEyXXksyo75o7kOH6S6sz6MvKfrLzfp2BA&s",
+      rating: 4.9,
+      prepTime: "40 minutes",
+      cookTime: "35 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Spiced chickpea curry cooked in a tomato-based gravy",
+      ingredients: [
+        { quantity: "2 cups", name: "chickpeas", notes: "soaked overnight" },
+        { quantity: "1", name: "onion", notes: "chopped" },
+        { quantity: "2", name: "tomatoes", notes: "pureed" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "cumin powder" },
+        { quantity: "1 teaspoon", name: "ginger garlic paste" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Pressure cook soaked chickpeas until tender.",
+        "Heat oil, sauté onion and ginger garlic paste until golden.",
+        "Add tomato puree, spices, and cook until oil separates.",
+        "Add chickpeas and simmer for 15 minutes.",
+        "Serve hot with rice or bhature.",
+      ],
+      nutritionalInfo: {
+        calories: "310",
+        fat: "9g",
+        protein: "13g",
+        carbs: "48g",
+      },
+      dietaryInfo: ["Vegetarian", "Vegan", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chickpea",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 30,
+      title: "Rajma (Kidney Bean Curry)",
+      image:
+        "https://www.indianveggiedelight.com/wp-content/uploads/2022/10/instant-pot-rajma-masala.jpg",
+      rating: 4.8,
+      prepTime: "45 minutes",
+      cookTime: "40 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Slow-cooked kidney beans in a spiced tomato curry",
+      ingredients: [
+        { quantity: "2 cups", name: "kidney beans", notes: "soaked overnight" },
+        { quantity: "1", name: "onion", notes: "chopped" },
+        { quantity: "2", name: "tomatoes", notes: "pureed" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "coriander powder" },
+        { quantity: "1 teaspoon", name: "ginger garlic paste" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Pressure cook soaked rajma until soft.",
+        "Heat oil, sauté onion and ginger garlic paste until golden.",
+        "Add tomato puree, spices, and cook until oil separates.",
+        "Add rajma with some water and simmer until thickened.",
+        "Serve hot with steamed basmati rice.",
+      ],
+      nutritionalInfo: {
+        calories: "330",
+        fat: "8g",
+        protein: "14g",
+        carbs: "52g",
+      },
+      dietaryInfo: ["Vegetarian", "Vegan", "Gluten-Free", "High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Rajma",
+      cookingMethod: "Stovetop",
+    },
+  ],
+  dinner: [
+    {
+      id: 31,
+      title: "Palak Paneer",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQzInsg3XxMcqrU233PO3em4HBiNRWuDwMWoQ&s",
+      rating: 4.9,
+      prepTime: "35 minutes",
+      cookTime: "20 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Soft paneer cubes cooked in a creamy spinach gravy",
+      ingredients: [
+        { quantity: "250g", name: "paneer", notes: "cubed" },
+        { quantity: "2 cups", name: "spinach", notes: "blanched and pureed" },
+        { quantity: "1", name: "onion", notes: "finely chopped" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "ginger garlic paste" },
+        { quantity: "1/2 cup", name: "cream" },
+        { quantity: "1 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Blanch spinach, cool, and blend to a smooth puree.",
+        "Heat oil in a pan, sauté onions until golden.",
+        "Add ginger garlic paste and cook for a minute.",
+        "Stir in spinach puree, salt, and garam masala.",
+        "Add paneer cubes and simmer for 5–7 minutes.",
+        "Finish with cream and mix gently.",
+        "Serve hot with naan or rice.",
+      ],
+      nutritionalInfo: {
+        calories: "280",
+        fat: "18g",
+        protein: "14g",
+        carbs: "16g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Dinner", "Lunch"],
+      mainIngredient: "Paneer",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 32,
+      title: "Chicken Curry",
+      image:
+        "https://www.cubesnjuliennes.com/wp-content/uploads/2020/07/Instant-Pot-Chicken-Curry-Recipe-500x500.jpg",
+      rating: 4.8,
+      prepTime: "40 minutes",
+      cookTime: "35 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Spicy and flavorful Indian-style chicken curry",
+      ingredients: [
+        { quantity: "500g", name: "chicken", notes: "bone-in pieces" },
+        { quantity: "2", name: "onions", notes: "finely chopped" },
+        { quantity: "2", name: "tomatoes", notes: "pureed" },
+        { quantity: "1 teaspoon", name: "turmeric powder" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "ginger garlic paste" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Heat oil, sauté onions until golden.",
+        "Add ginger garlic paste and cook for a minute.",
+        "Stir in tomatoes, turmeric, and salt; cook until oil separates.",
+        "Add chicken pieces and cook until sealed.",
+        "Pour water as needed, cover, and simmer for 25 minutes.",
+        "Finish with garam masala and cook for 2 minutes.",
+        "Serve hot with steamed rice or roti.",
+      ],
+      nutritionalInfo: {
+        calories: "350",
+        fat: "20g",
+        protein: "28g",
+        carbs: "12g",
+      },
+      dietaryInfo: ["Gluten-Free"],
+      mealType: ["Dinner", "Lunch"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 33,
+      title: "Aloo Gobi",
+      image:
+        "https://www.indianhealthyrecipes.com/wp-content/uploads/2022/03/aloo-gobi-recipe.jpg",
+      rating: 4.7,
+      prepTime: "30 minutes",
+      cookTime: "25 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Potatoes and cauliflower cooked with aromatic spices",
+      ingredients: [
+        { quantity: "2", name: "potatoes", notes: "cubed" },
+        { quantity: "1 cup", name: "cauliflower florets" },
+        { quantity: "1", name: "onion", notes: "chopped" },
+        { quantity: "1 teaspoon", name: "cumin seeds" },
+        { quantity: "1 teaspoon", name: "turmeric powder" },
+        { quantity: "1 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Heat oil in a pan, add cumin seeds and let them splutter.",
+        "Sauté onions until golden.",
+        "Add turmeric, chili powder, and salt; mix well.",
+        "Add potatoes and cauliflower, stir to coat with spices.",
+        "Cover and cook on low heat until vegetables are tender.",
+        "Serve hot with roti or paratha.",
+      ],
+      nutritionalInfo: {
+        calories: "220",
+        fat: "8g",
+        protein: "5g",
+        carbs: "34g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Dinner", "Lunch"],
+      mainIngredient: "Potato",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 34,
+      title: "Matar Paneer",
+      image:
+        "https://www.vegrecipesofindia.com/wp-content/uploads/2021/02/matar-paneer-2.jpg",
+      rating: 4.8,
+      prepTime: "35 minutes",
+      cookTime: "25 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description:
+        "Cottage cheese and green peas cooked in a spiced tomato gravy",
+      ingredients: [
+        { quantity: "250g", name: "paneer", notes: "cubed" },
+        { quantity: "1 cup", name: "green peas" },
+        { quantity: "2", name: "tomatoes", notes: "pureed" },
+        { quantity: "1", name: "onion", notes: "chopped" },
+        { quantity: "1 teaspoon", name: "cumin seeds" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Heat oil in a pan, add cumin seeds.",
+        "Sauté onions until golden brown.",
+        "Add tomato puree, salt, and spices; cook until oil separates.",
+        "Add green peas and some water, simmer until peas are tender.",
+        "Add paneer cubes and cook for 5 minutes.",
+        "Finish with garam masala and serve hot with naan or rice.",
+      ],
+      nutritionalInfo: {
+        calories: "290",
+        fat: "16g",
+        protein: "14g",
+        carbs: "22g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Dinner", "Lunch"],
+      mainIngredient: "Paneer",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 35,
+      title: "Baingan Bharta",
+      image:
+        "https://upload.wikimedia.org/wikipedia/commons/4/4d/Baigan_Bharta_from_Nagpur.JPG",
+      rating: 4.7,
+      prepTime: "30 minutes",
+      cookTime: "25 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Roasted eggplant mashed and cooked with spices",
+      ingredients: [
+        { quantity: "1 large", name: "eggplant", notes: "roasted and mashed" },
+        { quantity: "1", name: "onion", notes: "chopped" },
+        { quantity: "2", name: "tomatoes", notes: "chopped" },
+        { quantity: "1 teaspoon", name: "garlic", notes: "minced" },
+        { quantity: "1 teaspoon", name: "coriander powder" },
+        { quantity: "1 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Roast eggplant over flame until skin chars, peel and mash.",
+        "Heat oil, sauté onions until golden.",
+        "Add garlic and cook briefly.",
+        "Stir in tomatoes, chili powder, coriander powder, and salt.",
+        "Cook until tomatoes soften, then add mashed eggplant.",
+        "Mix well and cook for 5–7 minutes.",
+        "Serve hot with roti or paratha.",
+      ],
+      nutritionalInfo: {
+        calories: "180",
+        fat: "9g",
+        protein: "4g",
+        carbs: "24g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Dinner", "Lunch"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 36,
+      title: "Masoor Dal",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQVIbyKLDUHbXA_krLojlIhLlLJbJYktiNkRQ&s",
+      rating: 4.8,
+      prepTime: "30 minutes",
+      cookTime: "25 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      description: "Red lentils cooked with garlic and Indian spices",
+      ingredients: [
+        { quantity: "1 cup", name: "masoor dal (red lentils)" },
+        { quantity: "2 cups", name: "water" },
+        { quantity: "1 teaspoon", name: "turmeric powder" },
+        { quantity: "1 teaspoon", name: "mustard seeds" },
+        { quantity: "2 cloves", name: "garlic", notes: "minced" },
+        { quantity: "1 teaspoon", name: "ghee" },
+        { quantity: "1", name: "green chili", notes: "chopped" },
+      ],
+      instructions: [
+        "Rinse masoor dal and pressure cook with turmeric and water until soft.",
+        "Heat ghee in a pan, add mustard seeds and let them splutter.",
+        "Add garlic and green chili, sauté until fragrant.",
+        "Pour the tempering into cooked dal and simmer for 5 minutes.",
+        "Serve hot with rice or roti.",
+      ],
+      nutritionalInfo: {
+        calories: "210",
+        fat: "6g",
+        protein: "12g",
+        carbs: "30g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Dinner", "Lunch"],
+      mainIngredient: "Dal",
+      cookingMethod: "Stovetop",
+    },
+  ],
+  dessert: [
+    {
+      id: 37,
+      title: "Gulab Jamun",
+      image:
+        "https://static.toiimg.com/thumb/63799510.cms?width=1200&height=900",
+      rating: 4.9,
+      prepTime: "45 minutes",
+      cookTime: "20 minutes",
+      servings: 4,
+      description: "Soft, fried milk-based dumplings soaked in sugar syrup",
+      cuisine: "Indian",
+      category: "Dessert",
+      ingredients: [
+        { quantity: "1 cup", name: "milk powder" },
+        { quantity: "1/4 cup", name: "all-purpose flour" },
+        { quantity: "2 tablespoons", name: "ghee", notes: "melted" },
+        { quantity: "1/4 teaspoon", name: "baking soda" },
+        { quantity: "1/4 cup", name: "milk", notes: "warm" },
+        { quantity: "1 cup", name: "sugar" },
+        { quantity: "1/2 teaspoon", name: "cardamom powder" },
+      ],
+      instructions: [
+        "Mix milk powder, flour, baking soda, and ghee in a bowl.",
+        "Gradually add warm milk and knead into a soft dough.",
+        "Shape into small smooth balls.",
+        "Deep fry on low heat until golden brown.",
+        "Prepare sugar syrup with sugar, water, and cardamom powder.",
+        "Soak fried gulab jamuns in hot syrup for at least 1 hour before serving.",
+      ],
+      nutritionalInfo: {
+        calories: "300",
+        fat: "12g",
+        protein: "5g",
+        carbs: "40g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Dessert", "Snacks"],
+      mainIngredient: "All-purpose Flour",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 38,
+      title: "Rasgulla",
+      image:
+        "https://madhurasrecipe.com/wp-content/uploads/2023/10/Rasgulla-Featured-Image.jpg",
+      rating: 4.8,
+      prepTime: "50 minutes",
+      cookTime: "25 minutes",
+      servings: 5,
+      description: "Soft and spongy cheese balls soaked in light sugar syrup",
+      cuisine: "Indian",
+      category: "Dessert",
+      ingredients: [
+        { quantity: "1 liter", name: "milk" },
+        { quantity: "2 tablespoons", name: "lemon juice" },
+        { quantity: "1 cup", name: "sugar" },
+        { quantity: "4 cups", name: "water" },
+        { quantity: "1/2 teaspoon", name: "cardamom powder" },
+      ],
+      instructions: [
+        "Boil milk and curdle it using lemon juice to make chenna.",
+        "Strain and knead chenna until smooth.",
+        "Divide into small balls without cracks.",
+        "Boil sugar and water to prepare syrup.",
+        "Add rasgulla balls and cook covered until they double in size.",
+        "Cool and serve in syrup.",
+      ],
+      nutritionalInfo: {
+        calories: "250",
+        fat: "6g",
+        protein: "7g",
+        carbs: "42g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Dessert", "Snacks"],
+      mainIngredient: "All-purpose Flour",
+      cookingMethod: "Boiling",
+    },
+    {
+      id: 39,
+      title: "Kheer",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQWq5RkgNtHG4LL98GYxvZaCmwn8gXF5eFl5Q&s",
+      rating: 4.9,
+      prepTime: "40 minutes",
+      cookTime: "40 minutes",
+      servings: 4,
+      description: "Traditional Indian rice pudding with nuts and saffron",
+      cuisine: "Indian",
+      category: "Dessert",
+      ingredients: [
+        { quantity: "1/4 cup", name: "basmati rice", notes: "soaked" },
+        { quantity: "1 liter", name: "milk" },
+        { quantity: "1/2 cup", name: "sugar" },
+        { quantity: "1/4 teaspoon", name: "cardamom powder" },
+        { quantity: "10", name: "almonds", notes: "sliced" },
+        { quantity: "10", name: "cashews", notes: "chopped" },
+        { quantity: "1 pinch", name: "saffron strands" },
+      ],
+      instructions: [
+        "Boil milk in a heavy pan and add soaked rice.",
+        "Simmer on low heat until rice is cooked and mixture thickens.",
+        "Add sugar, cardamom, saffron, and stir well.",
+        "Garnish with almonds and cashews.",
+        "Serve warm or chilled.",
+      ],
+      nutritionalInfo: {
+        calories: "280",
+        fat: "10g",
+        protein: "8g",
+        carbs: "40g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Dessert"],
+      mainIngredient: "Milk",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 40,
+      title: "Jalebi",
+      image:
+        "https://www.chilitochoc.com/wp-content/uploads/2025/03/instant-jalebi-recipe.jpg",
+      rating: 4.7,
+      prepTime: "1 hour",
+      cookTime: "25 minutes",
+      servings: 6,
+      description: "Crispy, deep-fried spirals soaked in sugar syrup",
+      cuisine: "Indian",
+      category: "Dessert",
+      ingredients: [
+        { quantity: "1 cup", name: "all-purpose flour" },
+        { quantity: "1 tablespoon", name: "cornflour" },
+        { quantity: "1/2 teaspoon", name: "baking powder" },
+        { quantity: "1 cup", name: "yogurt" },
+        { quantity: "1 cup", name: "sugar" },
+        { quantity: "1/2 teaspoon", name: "cardamom powder" },
+        { quantity: "1 teaspoon", name: "lemon juice" },
+      ],
+      instructions: [
+        "Mix flour, cornflour, baking powder, and yogurt into a smooth batter.",
+        "Ferment batter for 6-8 hours or until slightly sour.",
+        "Prepare sugar syrup with sugar, water, cardamom, and lemon juice.",
+        "Pipe batter into hot oil in spiral shapes.",
+        "Fry until golden and crisp, then dip in warm syrup.",
+        "Serve hot for best taste.",
+      ],
+      nutritionalInfo: {
+        calories: "310",
+        fat: "12g",
+        protein: "5g",
+        carbs: "45g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Dessert", "Snacks"],
+      mainIngredient: "All-purpose Flour",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 41,
+      title: "Besan Ladoo",
+      image:
+        "https://cdn1.foodviva.com/static-content/food-images/desserts-sweets-recipes/besan-laddu/besan-laddu.jpg",
+      rating: 4.8,
+      prepTime: "30 minutes",
+      cookTime: "20 minutes",
+      servings: 6,
+      description: "Sweet gram flour balls flavored with ghee and cardamom",
+      cuisine: "Indian",
+      category: "Dessert",
+      ingredients: [
+        { quantity: "2 cups", name: "besan (gram flour)" },
+        { quantity: "1/2 cup", name: "ghee" },
+        { quantity: "3/4 cup", name: "powdered sugar" },
+        { quantity: "1/2 teaspoon", name: "cardamom powder" },
+        { quantity: "10", name: "almonds", notes: "chopped" },
+      ],
+      instructions: [
+        "Roast besan in ghee on low flame until golden and aromatic.",
+        "Let mixture cool slightly before adding powdered sugar.",
+        "Mix in cardamom powder and almonds.",
+        "Shape into round ladoos while warm.",
+        "Cool completely before storing.",
+      ],
+      nutritionalInfo: {
+        calories: "260",
+        fat: "11g",
+        protein: "6g",
+        carbs: "34g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Dessert", "Snacks"],
+      mainIngredient: "Besan",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 42,
+      title: "Gajar Ka Halwa",
+      image:
+        "https://www.cubesnjuliennes.com/wp-content/uploads/2019/11/Gajar-Ka-Halwa-2.jpg",
+      rating: 4.9,
+      prepTime: "45 minutes",
+      cookTime: "45 minutes",
+      servings: 5,
+      description: "Rich carrot pudding made with milk, ghee, and nuts",
+      cuisine: "Indian",
+      category: "Dessert",
+      ingredients: [
+        { quantity: "4 cups", name: "grated carrots" },
+        { quantity: "2 cups", name: "milk" },
+        { quantity: "1/2 cup", name: "sugar" },
+        { quantity: "1/4 cup", name: "ghee" },
+        { quantity: "1/2 teaspoon", name: "cardamom powder" },
+        { quantity: "10", name: "cashews", notes: "chopped" },
+        { quantity: "10", name: "raisins" },
+      ],
+      instructions: [
+        "Heat ghee in a pan and sauté grated carrots for a few minutes.",
+        "Add milk and simmer until carrots are cooked and milk reduces.",
+        "Stir in sugar and cook until mixture thickens.",
+        "Add cardamom, cashews, and raisins.",
+        "Serve warm or chilled.",
+      ],
+      nutritionalInfo: {
+        calories: "290",
+        fat: "12g",
+        protein: "6g",
+        carbs: "40g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Dessert"],
+      mainIngredient: "All-purpose Flour",
+      cookingMethod: "Stovetop",
+    },
+  ],
+  appetizers: [
+    {
+      id: 43,
+      title: "Samosa",
+      image:
+        "https://static.toiimg.com/thumb/61050397.cms?imgsize=246859&width=800&height=800",
+      rating: 4.9,
+      prepTime: "45 minutes",
+      cookTime: "25 minutes",
+      servings: 6,
+      description:
+        "Crispy, deep-fried pastry filled with spiced potatoes and peas",
+      cuisine: "Indian",
+      category: "Snack",
+      ingredients: [
+        { quantity: "2 cups", name: "all-purpose flour" },
+        { quantity: "4 tablespoons", name: "oil" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "3", name: "potatoes", notes: "boiled and mashed" },
+        { quantity: "1/2 cup", name: "peas" },
+        { quantity: "1 teaspoon", name: "cumin seeds" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+      ],
+      instructions: [
+        "Mix flour, salt, and oil. Add water gradually to knead a firm dough.",
+        "Heat oil and sauté cumin seeds, peas, and mashed potatoes.",
+        "Add garam masala and cook filling for 2–3 minutes.",
+        "Roll dough into discs, cut in half, and shape into cones.",
+        "Fill cones with potato mixture and seal edges.",
+        "Deep fry until golden and crisp.",
+      ],
+      nutritionalInfo: {
+        calories: "250",
+        fat: "10g",
+        protein: "5g",
+        carbs: "35g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Appetizer", "Snack"],
+      mainIngredient: "Potato",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 44,
+      title: "Paneer Tikka",
+      image: "https://orders.popskitchen.in/storage/2024/09/image-358.png",
+      rating: 4.8,
+      prepTime: "30 minutes",
+      cookTime: "20 minutes",
+      servings: 4,
+      description: "Grilled spiced paneer cubes with bell peppers and onions",
+      cuisine: "Indian",
+      category: "Appetizer",
+      ingredients: [
+        { quantity: "250 grams", name: "paneer", notes: "cubed" },
+        { quantity: "1/2 cup", name: "yogurt" },
+        { quantity: "1 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "lemon juice" },
+        { quantity: "1 tablespoon", name: "ginger-garlic paste" },
+        { quantity: "1", name: "bell pepper", notes: "cubed" },
+      ],
+      instructions: [
+        "Mix yogurt, spices, lemon juice, and ginger-garlic paste for marinade.",
+        "Coat paneer and bell pepper cubes. Marinate for 30 minutes.",
+        "Skewer alternately paneer and peppers.",
+        "Grill or bake until lightly charred and smoky.",
+        "Serve hot with mint chutney.",
+      ],
+      nutritionalInfo: {
+        calories: "280",
+        fat: "15g",
+        protein: "14g",
+        carbs: "12g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Appetizer"],
+      mainIngredient: "Paneer",
+      cookingMethod: "Grilling",
+    },
+    {
+      id: 45,
+      title: "Vegetable Pakora",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSqxHWAXstc2X1NLPKNstpaRJs-uU4Rb_ESmQ&s",
+      rating: 4.7,
+      prepTime: "20 minutes",
+      cookTime: "15 minutes",
+      servings: 4,
+      description: "Crispy deep-fried vegetable fritters made with gram flour",
+      cuisine: "Indian",
+      category: "Snack",
+      ingredients: [
+        { quantity: "1 cup", name: "gram flour (besan)" },
+        { quantity: "1", name: "potato", notes: "thinly sliced" },
+        { quantity: "1", name: "onion", notes: "sliced" },
+        { quantity: "1/2 teaspoon", name: "turmeric powder" },
+        { quantity: "1 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "cumin seeds" },
+        { quantity: "1/2 cup", name: "water" },
+      ],
+      instructions: [
+        "Mix gram flour with water and spices to form a smooth batter.",
+        "Dip potato and onion slices into batter.",
+        "Heat oil and deep fry until golden brown.",
+        "Drain on paper towels and serve hot with chutney.",
+      ],
+      nutritionalInfo: {
+        calories: "220",
+        fat: "9g",
+        protein: "6g",
+        carbs: "30g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Appetizer", "Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 46,
+      title: "Dahi Puri",
+      image:
+        "https://hogr.app/blog/wp-content/uploads/2022/10/dahi-puri-chat-is-indian-road-side-snack-item-which-is-especially-popular-state-maharashtra-india-scaled.webp",
+      rating: 4.9,
+      prepTime: "25 minutes",
+      cookTime: "5 minutes",
+      servings: 4,
+      description:
+        "Crispy puris filled with yogurt, chutneys, and spiced potatoes",
+      cuisine: "Indian",
+      category: "Chaat",
+      ingredients: [
+        { quantity: "10", name: "pani puris" },
+        { quantity: "1 cup", name: "yogurt", notes: "whisked" },
+        { quantity: "1", name: "potato", notes: "boiled and mashed" },
+        { quantity: "2 tablespoons", name: "tamarind chutney" },
+        { quantity: "2 tablespoons", name: "mint chutney" },
+        { quantity: "1/2 teaspoon", name: "chaat masala" },
+        { quantity: "1 tablespoon", name: "sev" },
+      ],
+      instructions: [
+        "Crack open puris from the top.",
+        "Fill each with mashed potatoes.",
+        "Top with yogurt, tamarind, and mint chutneys.",
+        "Sprinkle chaat masala and garnish with sev.",
+        "Serve immediately.",
+      ],
+      nutritionalInfo: {
+        calories: "180",
+        fat: "5g",
+        protein: "4g",
+        carbs: "28g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Appetizer", "Snack", "Chaat"],
+      mainIngredient: "Potato",
+      cookingMethod: "Assembling",
+    },
+    {
+      id: 47,
+      title: "Hara Bhara Kabab",
+      image:
+        "https://www.vegrecipesofindia.com/wp-content/uploads/2013/01/hara-bhara-kabab-recipe-1-500x500.jpg",
+      rating: 4.8,
+      prepTime: "30 minutes",
+      cookTime: "20 minutes",
+      servings: 4,
+      description: "Healthy pan-fried spinach and pea patties with spices",
+      cuisine: "Indian",
+      category: "Snack",
+      ingredients: [
+        { quantity: "1 cup", name: "spinach", notes: "blanched" },
+        { quantity: "1/2 cup", name: "peas", notes: "boiled" },
+        { quantity: "1", name: "potato", notes: "boiled and mashed" },
+        { quantity: "1 tablespoon", name: "ginger-garlic paste" },
+        { quantity: "1 teaspoon", name: "coriander powder" },
+        { quantity: "1 teaspoon", name: "green chili", notes: "chopped" },
+        { quantity: "2 tablespoons", name: "gram flour" },
+      ],
+      instructions: [
+        "Mash spinach, peas, and potato together.",
+        "Mix in spices, ginger-garlic paste, and gram flour.",
+        "Shape mixture into patties.",
+        "Pan-fry with a little oil until golden on both sides.",
+        "Serve hot with chutney.",
+      ],
+      nutritionalInfo: {
+        calories: "200",
+        fat: "6g",
+        protein: "7g",
+        carbs: "30g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Appetizer", "Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Pan-frying",
+    },
+    {
+      id: 48,
+      title: "Masala Papad",
+      image:
+        "https://www.whiskaffair.com/wp-content/uploads/2020/09/Masala-Papad-2-3.jpg",
+      rating: 4.7,
+      prepTime: "10 minutes",
+      cookTime: "5 minutes",
+      servings: 2,
+      description: "Crispy roasted papad topped with spicy onions and tomatoes",
+      cuisine: "Indian",
+      category: "Snack",
+      ingredients: [
+        { quantity: "2", name: "papads", notes: "roasted" },
+        { quantity: "1", name: "onion", notes: "finely chopped" },
+        { quantity: "1", name: "tomato", notes: "finely chopped" },
+        { quantity: "1/2 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "lemon juice" },
+        {
+          quantity: "1 tablespoon",
+          name: "coriander leaves",
+          notes: "chopped",
+        },
+      ],
+      instructions: [
+        "Roast papads until crisp.",
+        "Top with onion and tomato.",
+        "Sprinkle chili powder and drizzle lemon juice.",
+        "Garnish with coriander leaves.",
+        "Serve immediately as a crunchy snack.",
+      ],
+      nutritionalInfo: {
+        calories: "120",
+        fat: "3g",
+        protein: "4g",
+        carbs: "18g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Appetizer", "Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Roasting",
+    },
+  ],
+  sides: [
+    {
+      id: 49,
+      title: "Garlic Naan",
+      image:
+        "https://www.vegrecipesofindia.com/wp-content/uploads/2022/12/garlic-naan-3.jpg",
+      rating: 4.9,
+      prepTime: "25 minutes",
+      cookTime: "15 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Bread",
+      description: "Soft and fluffy naan bread flavored with garlic and butter",
+      ingredients: [
+        { quantity: "2 cups", name: "all-purpose flour" },
+        { quantity: "1 teaspoon", name: "baking powder" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 cup", name: "yogurt" },
+        { quantity: "1/2 cup", name: "warm water" },
+        { quantity: "1 tablespoon", name: "garlic", notes: "minced" },
+        { quantity: "2 tablespoons", name: "butter", notes: "melted" },
+      ],
+      instructions: [
+        "Mix flour, baking powder, and salt in a bowl.",
+        "Add yogurt and warm water gradually to form a soft dough.",
+        "Knead the dough for 5–6 minutes until smooth.",
+        "Cover and rest for 1 hour.",
+        "Divide into balls, roll out, and sprinkle minced garlic.",
+        "Cook on a hot skillet until golden spots appear.",
+        "Brush with melted butter before serving.",
+      ],
+      nutritionalInfo: {
+        calories: "200",
+        fat: "6g",
+        protein: "5g",
+        carbs: "32g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Side", "Snack"],
+      mainIngredient: "All-purpose Flour",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 50,
+      title: "Jeera Rice",
+      image:
+        "https://www.indianhealthyrecipes.com/wp-content/uploads/2022/12/jeera-rice-recipe.webp",
+      rating: 4.8,
+      prepTime: "20 minutes",
+      cookTime: "15 minutes",
+      servings: 3,
+      cuisine: "Indian",
+      category: "Rice",
+      description: "Aromatic basmati rice flavored with cumin seeds",
+      ingredients: [
+        { quantity: "1 cup", name: "basmati rice", notes: "soaked" },
+        { quantity: "2 cups", name: "water" },
+        { quantity: "1 tablespoon", name: "ghee" },
+        { quantity: "1 teaspoon", name: "cumin seeds" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1", name: "bay leaf" },
+        { quantity: "1", name: "clove" },
+      ],
+      instructions: [
+        "Heat ghee in a pan and add cumin seeds, bay leaf, and clove.",
+        "Sauté until fragrant.",
+        "Add soaked rice and stir gently for 1 minute.",
+        "Pour in water and add salt.",
+        "Cook covered on low heat until rice is fluffy.",
+        "Fluff with a fork and serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: "230",
+        fat: "5g",
+        protein: "4g",
+        carbs: "42g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Side"],
+      mainIngredient: "Rice",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 51,
+      title: "Boondi Raita",
+      image:
+        "https://dinedelicious.in/wp-content/uploads/2021/03/Boondi-raita-Recipe-2-500x500.jpg",
+      rating: 4.7,
+      prepTime: "10 minutes",
+      cookTime: "0 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Side",
+      description: "Refreshing yogurt dip with crispy boondi and spices",
+      ingredients: [
+        { quantity: "1 cup", name: "yogurt", notes: "whisked" },
+        { quantity: "1/2 cup", name: "boondi" },
+        { quantity: "1/2 teaspoon", name: "roasted cumin powder" },
+        { quantity: "1/4 teaspoon", name: "black salt" },
+        { quantity: "1/2 teaspoon", name: "sugar" },
+        {
+          quantity: "1 tablespoon",
+          name: "coriander leaves",
+          notes: "chopped",
+        },
+      ],
+      instructions: [
+        "Whisk yogurt in a bowl until smooth.",
+        "Add roasted cumin powder, black salt, and sugar. Mix well.",
+        "Soak boondi in warm water for 2 minutes, then squeeze out excess water.",
+        "Mix boondi into the yogurt.",
+        "Garnish with coriander leaves.",
+        "Serve chilled with biryani or paratha.",
+      ],
+      nutritionalInfo: {
+        calories: "120",
+        fat: "5g",
+        protein: "4g",
+        carbs: "15g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Side"],
+      mainIngredient: "Yogurt",
+      cookingMethod: "No-Cook",
+    },
+    {
+      id: 52,
+      title: "Mango Chutney",
+      image:
+        "https://sinfullyspicy.com/wp-content/uploads/2025/04/1200-by-1200-images-1-500x375.jpg",
+      rating: 4.8,
+      prepTime: "15 minutes",
+      cookTime: "20 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Condiment",
+      description: "Sweet and tangy mango chutney with Indian spices",
+      ingredients: [
+        { quantity: "2", name: "raw mangoes", notes: "peeled and chopped" },
+        { quantity: "1/2 cup", name: "sugar" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "mustard seeds" },
+        { quantity: "1 tablespoon", name: "vinegar" },
+        { quantity: "1/2 teaspoon", name: "cumin powder" },
+      ],
+      instructions: [
+        "Heat oil in a pan and add mustard seeds until they splutter.",
+        "Add chopped mangoes and cook for 5 minutes.",
+        "Mix in sugar, salt, chili powder, and cumin powder.",
+        "Simmer until mango softens and mixture thickens.",
+        "Add vinegar and cook for 2 more minutes.",
+        "Cool and store in an airtight jar.",
+      ],
+      nutritionalInfo: {
+        calories: "150",
+        fat: "1g",
+        protein: "1g",
+        carbs: "35g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Side", "Condiment"],
+      mainIngredient: "Fruit",
+      cookingMethod: "Stovetop",
+    },
+  ],
+  snacks: [
+    {
+      id: 53,
+      title: "Pakora",
+      image:
+        "https://i0.wp.com/www.happyandharried.com/wp-content/uploads/2017/10/ONION-PAKORA-ONION-FRITTERS-IN-CHICKPEA-FLOUR-BATTER-HAPPYHARRIED.jpg",
+      rating: 4.8,
+      prepTime: "25 minutes",
+      cookTime: "15 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Snack",
+      description: "Crunchy fritters made with chickpea flour and vegetables",
+      ingredients: [
+        { quantity: "1 cup", name: "chickpea flour (besan)" },
+        { quantity: "1", name: "onion", notes: "sliced" },
+        { quantity: "1", name: "potato", notes: "sliced" },
+        { quantity: "1 teaspoon", name: "cumin seeds" },
+        { quantity: "1/2 teaspoon", name: "turmeric powder" },
+        { quantity: "1/2 teaspoon", name: "red chili powder" },
+        { quantity: "as needed", name: "water" },
+        { quantity: "as needed", name: "oil", notes: "for deep frying" },
+      ],
+      instructions: [
+        "Mix chickpea flour, spices, and water to make a thick batter.",
+        "Dip onion and potato slices into the batter.",
+        "Heat oil and deep fry until golden and crispy.",
+        "Drain excess oil on paper towels.",
+        "Serve hot with chutney or tea.",
+      ],
+      nutritionalInfo: {
+        calories: "200",
+        fat: "10g",
+        protein: "5g",
+        carbs: "22g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 54,
+      title: "Aloo Tikki",
+      image:
+        "https://www.indianveggiedelight.com/wp-content/uploads/2023/07/aloo-tikki.jpg",
+      rating: 4.7,
+      prepTime: "30 minutes",
+      cookTime: "15 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Snack",
+      description: "Crispy potato patties with Indian spices",
+      ingredients: [
+        { quantity: "3", name: "potatoes", notes: "boiled and mashed" },
+        { quantity: "1/2 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1/2 teaspoon", name: "chaat masala" },
+        { quantity: "1/4 cup", name: "bread crumbs" },
+        { quantity: "2 tablespoons", name: "corn flour" },
+        { quantity: "as needed", name: "oil", notes: "for shallow frying" },
+      ],
+      instructions: [
+        "Combine mashed potatoes with spices, bread crumbs, and corn flour.",
+        "Shape mixture into flat patties.",
+        "Heat oil in a pan and shallow fry until golden and crispy on both sides.",
+        "Drain excess oil and serve hot with chutney.",
+      ],
+      nutritionalInfo: {
+        calories: "180",
+        fat: "8g",
+        protein: "4g",
+        carbs: "25g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free (if breadcrumbs are avoided)"],
+      mealType: ["Snack"],
+      mainIngredient: "Potato",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 55,
+      title: "Masala Corn",
+      image:
+        "https://i0.wp.com/eatsbyramya.com/wp-content/uploads/2022/04/IMG_7128_jpg.jpg?resize=1024%2C1020&ssl=1",
+      rating: 4.7,
+      prepTime: "10 minutes",
+      cookTime: "5 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Snack",
+      description: "Spicy and buttery sweet corn with lemon and chaat masala",
+      ingredients: [
+        { quantity: "1 cup", name: "sweet corn", notes: "boiled" },
+        { quantity: "1 teaspoon", name: "butter" },
+        { quantity: "1/2 teaspoon", name: "chaat masala" },
+        { quantity: "1/4 teaspoon", name: "red chili powder" },
+        { quantity: "1 teaspoon", name: "lemon juice" },
+        { quantity: "to taste", name: "salt" },
+      ],
+      instructions: [
+        "Melt butter in a pan.",
+        "Add boiled corn, spices, and lemon juice.",
+        "Toss well to coat.",
+        "Serve hot in cups.",
+      ],
+      nutritionalInfo: {
+        calories: "150",
+        fat: "4g",
+        protein: "5g",
+        carbs: "28g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 56,
+      title: "Bread Pizza",
+      image:
+        "https://www.awesomecuisine.com/wp-content/uploads/2009/05/bread-pizza.jpg",
+      rating: 4.8,
+      prepTime: "15 minutes",
+      cookTime: "10 minutes",
+      servings: 2,
+      cuisine: "Indian Fusion",
+      category: "Snack",
+      description:
+        "Quick and easy pizza made on bread slices with veggies and cheese",
+      ingredients: [
+        { quantity: "4", name: "bread slices" },
+        { quantity: "1/2 cup", name: "pizza sauce" },
+        { quantity: "1/2 cup", name: "mozzarella cheese", notes: "grated" },
+        { quantity: "1/2", name: "capsicum", notes: "chopped" },
+        { quantity: "1/2", name: "tomato", notes: "chopped" },
+        { quantity: "1/2 teaspoon", name: "oregano" },
+        { quantity: "1/4 teaspoon", name: "chili flakes" },
+      ],
+      instructions: [
+        "Spread pizza sauce on bread slices.",
+        "Top with chopped veggies and grated cheese.",
+        "Sprinkle oregano and chili flakes.",
+        "Toast in oven or pan until cheese melts.",
+        "Serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: "250",
+        fat: "9g",
+        protein: "10g",
+        carbs: "30g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Snack"],
+      mainIngredient: "Bread",
+      cookingMethod: "Baking",
+    },
+  ],
+  drinks: [
+    {
+      id: 57,
+      title: "Mango Lassi",
+      image:
+        "https://www.whiskaffair.com/wp-content/uploads/2020/05/Mango-Lassi-2-3.jpg",
+      rating: 4.9,
+      prepTime: "10 minutes",
+      cookTime: "0 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Beverage",
+      description: "Refreshing Indian mango yogurt smoothie",
+      ingredients: [
+        { quantity: "1 cup", name: "mango pulp" },
+        { quantity: "1 cup", name: "yogurt" },
+        { quantity: "1/2 cup", name: "milk" },
+        { quantity: "2 tablespoons", name: "sugar" },
+        { quantity: "1/4 teaspoon", name: "cardamom powder" },
+        { quantity: "4-5", name: "ice cubes" },
+      ],
+      instructions: [
+        "Combine mango pulp, yogurt, milk, sugar, and cardamom powder in a blender.",
+        "Blend until smooth and creamy.",
+        "Add ice cubes and blend again briefly.",
+        "Pour into glasses and serve chilled.",
+      ],
+      nutritionalInfo: {
+        calories: "180",
+        fat: "5g",
+        protein: "6g",
+        carbs: "30g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Drinks"],
+      mainIngredient: "Fruit",
+      cookingMethod: "Blending",
+    },
+    {
+      id: 58,
+      title: "Masala Chai",
+      image:
+        "https://www.teacupsfull.com/cdn/shop/articles/Screenshot_2023-10-20_at_11.07.13_AM.png?v=1697780292&width=749",
+      rating: 4.8,
+      prepTime: "15 minutes",
+      cookTime: "10 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Beverage",
+      description: "Spiced Indian tea made with milk and aromatic spices",
+      ingredients: [
+        { quantity: "2 cups", name: "water" },
+        { quantity: "1 cup", name: "milk" },
+        { quantity: "2 teaspoons", name: "black tea leaves" },
+        { quantity: "1", name: "cinnamon stick" },
+        { quantity: "2", name: "cardamom pods", notes: "crushed" },
+        { quantity: "1", name: "clove" },
+        { quantity: "1/2 inch", name: "ginger", notes: "grated" },
+        { quantity: "2 teaspoons", name: "sugar" },
+      ],
+      instructions: [
+        "Bring water to a boil in a saucepan.",
+        "Add cinnamon, cardamom, clove, and grated ginger. Boil for 2-3 minutes.",
+        "Add tea leaves and let simmer for 2 minutes.",
+        "Pour in milk and bring to a gentle boil.",
+        "Add sugar and stir well.",
+        "Strain into cups and serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: "90",
+        fat: "3g",
+        protein: "2g",
+        carbs: "14g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Drinks"],
+      mainIngredient: "Milk",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 59,
+      title: "Jaljeera",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ_8mmIhhP7ihBxDM7nUZqOePdvfITQd0YQ1g&s",
+      rating: 4.7,
+      prepTime: "10 minutes",
+      cookTime: "0 minutes",
+      servings: 3,
+      cuisine: "Indian",
+      category: "Beverage",
+      description: "Tangy and spicy Indian cumin-flavored drink",
+      ingredients: [
+        { quantity: "2 cups", name: "water" },
+        { quantity: "1/2 cup", name: "mint leaves" },
+        { quantity: "1/2 teaspoon", name: "cumin powder" },
+        { quantity: "1/2 teaspoon", name: "black salt" },
+        { quantity: "1/4 teaspoon", name: "chili powder" },
+        { quantity: "2 tablespoons", name: "lemon juice" },
+        { quantity: "4-5", name: "ice cubes" },
+      ],
+      instructions: [
+        "Blend mint leaves with a little water to make a smooth paste.",
+        "In a jug, combine water, mint paste, cumin powder, black salt, chili powder, and lemon juice.",
+        "Mix well and adjust seasoning if needed.",
+        "Add ice cubes and serve chilled.",
+      ],
+      nutritionalInfo: {
+        calories: "35",
+        fat: "0g",
+        protein: "1g",
+        carbs: "8g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Drinks"],
+      mainIngredient: "Water",
+      cookingMethod: "Blending",
+    },
+    {
+      id: 60,
+      title: "Cold Coffee",
+      image:
+        "https://www.milkmaid.in/sites/default/files/2024-05/Cold-Coffee-335x300.jpg",
+      rating: 4.8,
+      prepTime: "5 minutes",
+      cookTime: "0 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Beverage",
+      description: "Chilled coffee blended with milk and sugar",
+      ingredients: [
+        { quantity: "1 cup", name: "milk" },
+        { quantity: "1 tablespoon", name: "instant coffee" },
+        { quantity: "2 tablespoons", name: "sugar" },
+        { quantity: "4-5", name: "ice cubes" },
+        { quantity: "as needed", name: "whipped cream", notes: "optional" },
+        { quantity: "1 teaspoon", name: "chocolate syrup", notes: "optional" },
+      ],
+      instructions: [
+        "In a blender, combine milk, instant coffee, sugar, and ice cubes.",
+        "Blend until frothy and smooth.",
+        "Pour into glasses.",
+        "Top with whipped cream and drizzle chocolate syrup if desired.",
+        "Serve chilled.",
+      ],
+      nutritionalInfo: {
+        calories: "120",
+        fat: "3g",
+        protein: "4g",
+        carbs: "20g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Drinks"],
+      mainIngredient: "Milk",
+      cookingMethod: "Blending",
+    },
+    {
+      id: 61,
+      title: "Aam Panna",
+      image:
+        "https://i0.wp.com/binjalsvegkitchen.com/wp-content/uploads/2016/05/Pudina-Aam-Panna-H1-1.jpg?resize=600%2C900&ssl=1",
+      rating: 4.7,
+      prepTime: "20 minutes",
+      cookTime: "10 minutes",
+      servings: 3,
+      cuisine: "Indian",
+      category: "Beverage",
+      description: "Refreshing summer drink made with raw mango and spices",
+      ingredients: [
+        { quantity: "2", name: "raw mangoes" },
+        { quantity: "2 tablespoons", name: "sugar" },
+        { quantity: "1 teaspoon", name: "black salt" },
+        { quantity: "1 teaspoon", name: "cumin powder" },
+        { quantity: "2 tablespoons", name: "mint leaves", notes: "chopped" },
+        { quantity: "2 cups", name: "water" },
+        { quantity: "4-5", name: "ice cubes" },
+      ],
+      instructions: [
+        "Boil raw mangoes until soft. Let them cool.",
+        "Peel and extract pulp.",
+        "Blend mango pulp with sugar, black salt, cumin powder, and mint leaves.",
+        "Add water and mix well.",
+        "Serve with ice cubes.",
+      ],
+      nutritionalInfo: {
+        calories: "60",
+        fat: "0g",
+        protein: "1g",
+        carbs: "15g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Drinks"],
+      mainIngredient: "Fruit",
+      cookingMethod: "Blending",
+    },
+    {
+      id: 62,
+      title: "Thandai",
+      image:
+        "https://palatesdesire.com/wp-content/uploads/2022/03/Thandai-recipe@palates-desire-2.jpg",
+      rating: 4.9,
+      prepTime: "30 minutes",
+      cookTime: "0 minutes",
+      servings: 2,
+      cuisine: "Indian",
+      category: "Beverage",
+      description: "Traditional Indian spiced milk drink",
+      ingredients: [
+        { quantity: "2 cups", name: "milk" },
+        { quantity: "10", name: "almonds", notes: "soaked" },
+        { quantity: "5", name: "cashews", notes: "soaked" },
+        { quantity: "1 teaspoon", name: "poppy seeds" },
+        { quantity: "1 teaspoon", name: "fennel seeds" },
+        { quantity: "1/4 teaspoon", name: "cardamom powder" },
+        { quantity: "2 tablespoons", name: "sugar" },
+        { quantity: "4-5", name: "ice cubes" },
+      ],
+      instructions: [
+        "Grind soaked almonds, cashews, poppy seeds, fennel seeds, and cardamom into a fine paste.",
+        "Mix the paste with milk and sugar.",
+        "Stir well until blended.",
+        "Add ice cubes and serve chilled.",
+      ],
+      nutritionalInfo: {
+        calories: "200",
+        fat: "10g",
+        protein: "7g",
+        carbs: "20g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Drinks"],
+      mainIngredient: "Milk",
+      cookingMethod: "Blending",
+    },
+  ],
+  lowCarb: [
+    {
+      id: 63,
+      title: "Grilled Chicken Salad",
+      image:
+        "https://www.crunchycreamysweet.com/wp-content/uploads/2018/06/easy-grilled-chicken-salad-1.jpg",
+      rating: 4.8,
+      prepTime: "20 minutes",
+      cookTime: "10 minutes",
+      servings: 2,
+      cuisine: "American",
+      category: "Salad",
+      description:
+        "Fresh salad with grilled chicken, avocado, and a light dressing",
+      ingredients: [
+        { quantity: "1", name: "chicken breast", notes: "grilled and sliced" },
+        { quantity: "2 cups", name: "mixed greens" },
+        { quantity: "1/2", name: "avocado", notes: "sliced" },
+        { quantity: "1/4 cup", name: "cherry tomatoes", notes: "halved" },
+        { quantity: "2 tablespoons", name: "olive oil" },
+        { quantity: "1 tablespoon", name: "lemon juice" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Season and grill chicken breast until fully cooked, then slice.",
+        "In a large bowl, combine mixed greens, avocado slices, and cherry tomatoes.",
+        "Whisk olive oil, lemon juice, salt, and pepper to make dressing.",
+        "Drizzle dressing over the salad and toss lightly.",
+        "Top with grilled chicken slices and serve fresh.",
+      ],
+      nutritionalInfo: {
+        calories: "350",
+        fat: "20g",
+        protein: "30g",
+        carbs: "10g",
+      },
+      dietaryInfo: ["Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Grilling",
+      diet: "lowCarb",
+    },
+    {
+      id: 64,
+      title: "Zucchini Noodles with Pesto",
+      image:
+        "https://www.twopeasandtheirpod.com/wp-content/uploads/2013/08/Zucchini-Noodles-with-Pesto-7.jpg",
+      rating: 4.7,
+      prepTime: "15 minutes",
+      cookTime: "5 minutes",
+      servings: 2,
+      cuisine: "Italian",
+      category: "Main Course",
+      description:
+        "Light and flavorful zucchini noodles tossed in homemade pesto",
+      ingredients: [
+        { quantity: "2", name: "zucchini", notes: "spiralized" },
+        { quantity: "1/4 cup", name: "basil leaves" },
+        { quantity: "2 tablespoons", name: "parmesan cheese" },
+        { quantity: "1 tablespoon", name: "pine nuts" },
+        { quantity: "1 clove", name: "garlic" },
+        { quantity: "2 tablespoons", name: "olive oil" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+      ],
+      instructions: [
+        "Blend basil, parmesan, pine nuts, garlic, olive oil, salt, and pepper to make pesto.",
+        "Lightly sauté zucchini noodles in a pan for 2–3 minutes.",
+        "Toss zucchini noodles with pesto until evenly coated.",
+        "Serve immediately with extra parmesan if desired.",
+      ],
+      nutritionalInfo: {
+        calories: "210",
+        fat: "15g",
+        protein: "7g",
+        carbs: "10g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free", "Low-Carb"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Stovetop",
+      diet: "lowCarb",
+    },
+    {
+      id: 65,
+      title: "Cauliflower Fried Rice",
+      image:
+        "https://thebigmansworld.com/wp-content/uploads/2022/06/cauliflower-fried-rice-recipe.jpg",
+      rating: 4.9,
+      prepTime: "20 minutes",
+      cookTime: "10 minutes",
+      servings: 3,
+      cuisine: "Asian",
+      category: "Main Course",
+      description: "Low-carb alternative to fried rice using cauliflower",
+      ingredients: [
+        { quantity: "2 cups", name: "cauliflower rice" },
+        { quantity: "1/2 cup", name: "bell peppers", notes: "chopped" },
+        { quantity: "1/2 cup", name: "carrots", notes: "chopped" },
+        { quantity: "1", name: "egg", notes: "beaten" },
+        { quantity: "2 tablespoons", name: "soy sauce" },
+        { quantity: "1 teaspoon", name: "ginger", notes: "grated" },
+        { quantity: "1 clove", name: "garlic", notes: "minced" },
+        { quantity: "2 tablespoons", name: "olive oil" },
+      ],
+      instructions: [
+        "Heat olive oil in a pan and sauté garlic and ginger.",
+        "Add carrots and bell peppers; cook until slightly tender.",
+        "Push veggies aside, scramble the beaten egg, then mix in.",
+        "Add cauliflower rice and stir-fry for 5 minutes.",
+        "Pour in soy sauce and toss everything together before serving.",
+      ],
+      nutritionalInfo: {
+        calories: "150",
+        fat: "8g",
+        protein: "6g",
+        carbs: "12g",
+      },
+      dietaryInfo: ["Low-Carb", "Gluten-Free (if using tamari)", "Dairy-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Stovetop",
+      diet: "lowCarb",
+    },
+    {
+      id: 66,
+      title: "Keto Avocado Egg Salad",
+      image: "https://i.ytimg.com/vi/KCZJsIkEmyw/maxresdefault.jpg",
+      rating: 4.8,
+      prepTime: "10 minutes",
+      cookTime: "0 minutes",
+      servings: 2,
+      cuisine: "American",
+      category: "Salad",
+      description: "Creamy avocado and egg salad with a hint of lime",
+      ingredients: [
+        { quantity: "2", name: "hard-boiled eggs", notes: "chopped" },
+        { quantity: "1", name: "avocado", notes: "mashed" },
+        { quantity: "1 tablespoon", name: "mayonnaise" },
+        { quantity: "1 teaspoon", name: "lime juice" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/4 teaspoon", name: "black pepper" },
+        { quantity: "1 tablespoon", name: "chopped parsley" },
+      ],
+      instructions: [
+        "Mash avocado in a mixing bowl.",
+        "Add chopped boiled eggs, mayonnaise, lime juice, salt, and pepper.",
+        "Mix well until creamy.",
+        "Garnish with parsley and serve chilled or as a sandwich filling.",
+      ],
+      nutritionalInfo: {
+        calories: "280",
+        fat: "24g",
+        protein: "10g",
+        carbs: "6g",
+      },
+      dietaryInfo: ["Keto", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Egg",
+      cookingMethod: "No-Cook",
+      diet: "lowCarb",
+    },
+    {
+      id: 67,
+      title: "Garlic Butter Shrimp",
+      image:
+        "https://www.jocooks.com/wp-content/uploads/2021/09/garlic-butter-shrimp-1-10.jpg",
+      rating: 4.9,
+      prepTime: "15 minutes",
+      cookTime: "10 minutes",
+      servings: 3,
+      cuisine: "Seafood",
+      category: "Main Course",
+      description: "Juicy shrimp cooked in garlic butter and lemon juice",
+      ingredients: [
+        { quantity: "1 lb", name: "shrimp", notes: "peeled and deveined" },
+        { quantity: "2 tablespoons", name: "butter" },
+        { quantity: "3 cloves", name: "garlic", notes: "minced" },
+        { quantity: "1 tablespoon", name: "lemon juice" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/4 teaspoon", name: "black pepper" },
+        { quantity: "1 tablespoon", name: "chopped parsley" },
+      ],
+      instructions: [
+        "Melt butter in a skillet over medium heat.",
+        "Add minced garlic and sauté until fragrant.",
+        "Add shrimp, salt, and pepper; cook for 2–3 minutes per side until pink.",
+        "Stir in lemon juice and garnish with parsley.",
+        "Serve hot with rice or bread.",
+      ],
+      nutritionalInfo: {
+        calories: "220",
+        fat: "11g",
+        protein: "25g",
+        carbs: "2g",
+      },
+      dietaryInfo: ["Gluten-Free", "Low-Carb"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Fish",
+      cookingMethod: "Stovetop",
+      diet: "lowCarb",
+    },
+    {
+      id: 68,
+      title: "Stuffed Bell Peppers",
+      image:
+        "https://embed.widencdn.net/img/beef/t9bwp7fitq/exact/Stuffed%20Peppers%20-%20NCBA%20Beef%20Aug%20202431717.jpg?keep=c&u=7fueml",
+      rating: 4.7,
+      prepTime: "30 minutes",
+      cookTime: "25 minutes",
+      servings: 4,
+      cuisine: "American",
+      category: "Main Course",
+      description:
+        "Bell peppers stuffed with a flavorful mix of ground meat and vegetables",
+      ingredients: [
+        { quantity: "3", name: "bell peppers", notes: "halved and deseeded" },
+        { quantity: "1/2 lb", name: "ground turkey" },
+        { quantity: "1/2 cup", name: "chopped spinach" },
+        { quantity: "1/4 cup", name: "cheese", notes: "grated" },
+        { quantity: "1 teaspoon", name: "garlic powder" },
+        { quantity: "1 teaspoon", name: "paprika" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+      ],
+      instructions: [
+        "Preheat oven to 375°F (190°C).",
+        "Cook ground turkey in a skillet with garlic powder, paprika, salt, and pepper until browned.",
+        "Mix in chopped spinach and cook for another 2 minutes.",
+        "Stuff the turkey mixture into halved bell peppers.",
+        "Top with grated cheese and bake for 20–25 minutes.",
+        "Serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: "280",
+        fat: "14g",
+        protein: "25g",
+        carbs: "12g",
+      },
+      dietaryInfo: ["Gluten-Free", "Low-Carb"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Baking",
+      diet: "lowCarb",
+    },
+  ],
+  paleo: [
+    {
+      id: 69,
+      title: "Grilled Chicken with Roasted Vegetables",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS6OQCqpilQ0KIOxPclC6Z75ANJhNk0FGXvpQ&s",
+      rating: 4.9,
+      prepTime: "30 minutes",
+      cookTime: "20 minutes",
+      servings: 2,
+      description: "Juicy grilled chicken with a side of roasted veggies",
+      cuisine: "American",
+      category: "Main Course",
+      ingredients: [
+        { quantity: "2", name: "chicken breasts", notes: "boneless, skinless" },
+        { quantity: "2 tablespoons", name: "olive oil" },
+        { quantity: "1 teaspoon", name: "garlic powder" },
+        { quantity: "1 teaspoon", name: "paprika" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+        {
+          quantity: "1 cup",
+          name: "mixed vegetables",
+          notes: "zucchini, bell peppers, carrots",
+        },
+      ],
+      instructions: [
+        "Preheat the grill to medium-high heat.",
+        "Season chicken breasts with olive oil, garlic powder, paprika, salt, and pepper.",
+        "Grill chicken for 6-7 minutes per side until fully cooked.",
+        "Toss vegetables with olive oil, salt, and pepper, then roast at 200°C (400°F) for 20 minutes.",
+        "Serve grilled chicken with roasted vegetables on the side.",
+      ],
+      nutritionalInfo: {
+        calories: "420",
+        fat: "18g",
+        protein: "46g",
+        carbs: "15g",
+      },
+      dietaryInfo: ["Gluten-Free", "High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Grilling",
+      diet: "paleo",
+    },
+    {
+      id: 70,
+      title: "Zucchini Noodles with Meatballs",
+      image:
+        "https://lifemadesweeter.com/wp-content/uploads/Zucchini-Noodles-with-Meatballs-recipe-photo-picture-1-10-500x375.jpg",
+      rating: 4.9,
+      prepTime: "35 minutes",
+      cookTime: "25 minutes",
+      servings: 3,
+      description:
+        "Savory meatballs served with zucchini noodles and marinara sauce",
+      cuisine: "Italian",
+      category: "Main Course",
+      ingredients: [
+        { quantity: "2", name: "zucchinis", notes: "spiralized" },
+        { quantity: "1/2 pound", name: "ground beef" },
+        { quantity: "1", name: "egg" },
+        { quantity: "1/4 cup", name: "almond flour" },
+        { quantity: "1 teaspoon", name: "garlic powder" },
+        { quantity: "1 teaspoon", name: "oregano" },
+        { quantity: "1 cup", name: "marinara sauce", notes: "sugar-free" },
+      ],
+      instructions: [
+        "Preheat oven to 375°F (190°C).",
+        "Mix ground beef, egg, almond flour, garlic powder, and oregano.",
+        "Form into small meatballs and bake for 15-20 minutes until cooked.",
+        "In a pan, heat marinara sauce and add cooked meatballs.",
+        "Lightly sauté zucchini noodles for 2-3 minutes until just tender.",
+        "Serve meatballs over zucchini noodles with sauce.",
+      ],
+      nutritionalInfo: {
+        calories: "390",
+        fat: "22g",
+        protein: "30g",
+        carbs: "12g",
+      },
+      dietaryInfo: ["Gluten-Free", "Low-Carb"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Beef",
+      cookingMethod: "Baking",
+      diet: "paleo",
+    },
+    {
+      id: 71,
+      title: "Baked Salmon with Lemon and Dill",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSGAbEDRmsrtFw3dfuEynddW2OVsb2WjbOkeg&s",
+      rating: 5.0,
+      prepTime: "25 minutes",
+      cookTime: "20 minutes",
+      servings: 2,
+      description: "Oven-baked salmon fillets with fresh lemon and dill",
+      cuisine: "Mediterranean",
+      category: "Main Course",
+      ingredients: [
+        { quantity: "2", name: "salmon fillets" },
+        { quantity: "1 tablespoon", name: "olive oil" },
+        { quantity: "1 teaspoon", name: "lemon juice" },
+        { quantity: "1 teaspoon", name: "dill", notes: "fresh, chopped" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+        { quantity: "1", name: "lemon", notes: "sliced for garnish" },
+      ],
+      instructions: [
+        "Preheat oven to 375°F (190°C).",
+        "Brush salmon fillets with olive oil, lemon juice, dill, salt, and pepper.",
+        "Place on a baking tray lined with parchment paper.",
+        "Bake for 15-20 minutes until salmon flakes easily with a fork.",
+        "Garnish with lemon slices and serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: "420",
+        fat: "26g",
+        protein: "40g",
+        carbs: "3g",
+      },
+      dietaryInfo: ["Gluten-Free", "Keto-Friendly"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Fish",
+      cookingMethod: "Baking",
+      diet: "paleo",
+    },
+    {
+      id: 72,
+      title: "Sweet Potato and Avocado Breakfast Hash",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ5b8E6eZfwLLJSOqkbwm6O4N0YozMgDlEofQ&s",
+      rating: 4.8,
+      prepTime: "30 minutes",
+      cookTime: "20 minutes",
+      servings: 2,
+      description: "A hearty sweet potato hash with avocado and eggs",
+      cuisine: "American",
+      category: "Breakfast",
+      ingredients: [
+        { quantity: "1", name: "sweet potato", notes: "diced" },
+        { quantity: "1/2 cup", name: "onion", notes: "chopped" },
+        { quantity: "1/2 cup", name: "bell peppers", notes: "chopped" },
+        { quantity: "1 tablespoon", name: "olive oil" },
+        { quantity: "2", name: "eggs" },
+        { quantity: "1", name: "avocado", notes: "sliced" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Heat olive oil in a skillet over medium heat.",
+        "Add diced sweet potatoes and cook until golden and tender.",
+        "Stir in onions and bell peppers; sauté until softened.",
+        "Make small wells and crack eggs into the pan, cooking until done.",
+        "Top with sliced avocado and serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: "370",
+        fat: "20g",
+        protein: "12g",
+        carbs: "34g",
+      },
+      dietaryInfo: ["Gluten-Free", "Vegetarian-Friendly"],
+      mealType: ["Breakfast"],
+      mainIngredient: "Potato",
+      cookingMethod: "Stovetop",
+      diet: "paleo",
+    },
+    {
+      id: 73,
+      title: "Coconut Chicken Curry",
+      image:
+        "https://hips.hearstapps.com/hmg-prod/images/190509-coconut-chicken-curry-157-1558039780.jpg?crop=0.8891228070175439xw:1xh;center,top&resize=1200:*",
+      rating: 4.9,
+      prepTime: "35 minutes",
+      cookTime: "30 minutes",
+      servings: 4,
+      description: "A rich and creamy coconut curry with chicken and spices",
+      cuisine: "Indian",
+      category: "Main Course",
+      ingredients: [
+        { quantity: "2", name: "chicken breasts", notes: "cubed" },
+        { quantity: "1 cup", name: "coconut milk" },
+        { quantity: "1/2 cup", name: "tomato puree" },
+        { quantity: "1 tablespoon", name: "curry powder" },
+        { quantity: "1/2 teaspoon", name: "turmeric" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1 tablespoon", name: "olive oil" },
+      ],
+      instructions: [
+        "Heat olive oil in a pan over medium heat.",
+        "Add chicken cubes and cook until lightly browned.",
+        "Stir in curry powder, turmeric, and salt.",
+        "Pour in tomato puree and cook for 5 minutes.",
+        "Add coconut milk and simmer for 15-20 minutes until chicken is tender.",
+        "Serve hot with rice or naan.",
+      ],
+      nutritionalInfo: {
+        calories: "480",
+        fat: "28g",
+        protein: "35g",
+        carbs: "18g",
+      },
+      dietaryInfo: ["Gluten-Free", "Dairy-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Stovetop",
+      diet: "paleo",
+    },
+  ],
+  keto: [
+    {
+      id: 74,
+      title: "Keto Butter Chicken",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSRSOBfEag9GHFrxriO65pEs5vA9HO_N0itnQ&s",
+      rating: 4.9,
+      description: "Creamy and rich butter chicken with a keto-friendly sauce",
+      prepTime: "30 minutes",
+      cookingTime: "25 minutes",
+      servings: 4,
+      cuisine: "Indian",
+      category: "Main Course",
+      ingredients: [
+        { quantity: "1 lb", name: "chicken breast", notes: "cubed" },
+        { quantity: "1/2 cup", name: "heavy cream" },
+        { quantity: "2 tablespoons", name: "butter" },
+        { quantity: "1/2 cup", name: "tomato puree" },
+        { quantity: "1 teaspoon", name: "garam masala" },
+        { quantity: "1 teaspoon", name: "turmeric" },
+        { quantity: "1 teaspoon", name: "garlic", notes: "minced" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Heat butter in a skillet and sauté garlic until fragrant.",
+        "Add chicken cubes and cook until browned.",
+        "Stir in tomato puree, garam masala, turmeric, and salt.",
+        "Simmer for 10 minutes, then stir in heavy cream.",
+        "Cook until sauce thickens and chicken is tender.",
+      ],
+      nutritionalInfo: {
+        calories: 320,
+        protein: "27g",
+        fat: "22g",
+        carbs: "5g",
+      },
+      dietaryInfo: ["Keto", "Low-Carb", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Stovetop",
+      diet: "keto",
+    },
+    {
+      id: 75,
+      title: "Keto Cauliflower Mac and Cheese",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRxr8TBUNTO6rrHitYox4Wfv9yxsCLVC9-xTg&s",
+      rating: 4.8,
+      description: "Cheesy and creamy mac and cheese made with cauliflower",
+      prepTime: "20 minutes",
+      cookingTime: "20 minutes",
+      servings: 4,
+      cuisine: "American",
+      category: "Side",
+      ingredients: [
+        { quantity: "2 cups", name: "cauliflower", notes: "chopped" },
+        { quantity: "1 cup", name: "cheddar cheese", notes: "shredded" },
+        { quantity: "1/2 cup", name: "heavy cream" },
+        { quantity: "2 tablespoons", name: "butter" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/4 teaspoon", name: "black pepper" },
+        { quantity: "1/2 teaspoon", name: "paprika" },
+      ],
+      instructions: [
+        "Steam cauliflower until tender.",
+        "In a pan, melt butter and stir in cream.",
+        "Add cheese, salt, pepper, and paprika to make sauce.",
+        "Mix cauliflower with cheese sauce until fully coated.",
+        "Serve hot as a keto mac and cheese substitute.",
+      ],
+      nutritionalInfo: {
+        calories: 280,
+        protein: "10g",
+        fat: "24g",
+        carbs: "6g",
+      },
+      dietaryInfo: ["Keto", "Low-Carb", "Vegetarian", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner", "Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Stovetop",
+      diet: "keto",
+    },
+    {
+      id: 76,
+      title: "Keto Avocado Bacon Salad",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRE2ccLXdd9BXwJd7bn1N7yl0G4VtD2FS_48w&s",
+      rating: 4.9,
+      description: "Fresh salad with crispy bacon and creamy avocado",
+      prepTime: "10 minutes",
+      cookingTime: "5 minutes",
+      servings: 2,
+      cuisine: "American",
+      category: "Salad",
+      ingredients: [
+        { quantity: "2", name: "avocados", notes: "diced" },
+        { quantity: "4 slices", name: "bacon", notes: "cooked and crumbled" },
+        { quantity: "2 cups", name: "lettuce", notes: "chopped" },
+        { quantity: "1/4 cup", name: "olive oil" },
+        { quantity: "1 tablespoon", name: "lemon juice" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+      ],
+      instructions: [
+        "Cook bacon until crispy, then crumble.",
+        "Toss lettuce, avocado, and bacon in a bowl.",
+        "Whisk olive oil, lemon juice, salt, and pepper for dressing.",
+        "Drizzle dressing over salad and serve fresh.",
+      ],
+      nutritionalInfo: {
+        calories: 350,
+        protein: "9g",
+        fat: "32g",
+        carbs: "8g",
+      },
+      dietaryInfo: ["Keto", "Low-Carb", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner", "Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "No-Cook",
+      diet: "keto",
+    },
+    {
+      id: 77,
+      title: "Keto Garlic Butter Steak",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRU7nmUNRez4x9qQ2psMchj2NvST7IJ43Ksag&s",
+      rating: 5.0,
+      description: "Juicy steak cooked in garlic butter for maximum flavor",
+      prepTime: "20 minutes",
+      cookingTime: "15 minutes",
+      servings: 2,
+      cuisine: "American",
+      category: "Main Course",
+      ingredients: [
+        { quantity: "1 lb", name: "ribeye steak" },
+        { quantity: "3 tablespoons", name: "butter" },
+        { quantity: "3 cloves", name: "garlic", notes: "minced" },
+        { quantity: "1 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+        { quantity: "1 teaspoon", name: "fresh rosemary" },
+      ],
+      instructions: [
+        "Season steak with salt and pepper.",
+        "Heat skillet with butter and garlic.",
+        "Cook steak to desired doneness.",
+        "Add rosemary and baste with garlic butter.",
+        "Rest steak before serving.",
+      ],
+      nutritionalInfo: {
+        calories: 450,
+        protein: "35g",
+        fat: "36g",
+        carbs: "2g",
+      },
+      dietaryInfo: ["Keto", "Low-Carb", "Gluten-Free", "High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Beef",
+      cookingMethod: "Stovetop",
+      diet: "keto",
+    },
+    {
+      id: 78,
+      title: "Keto Zucchini Lasagna",
+      image:
+        "https://www.allrecipes.com/thmb/kuPeHUeeKvxJjV6pLel-AFi016M=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/261026_UltimateLowCarbZucchiniLasagne_MFS_182-1-2x1-dec427f468f24181b1c2778ccae13735.jpg",
+      rating: 4.7,
+      description: "Delicious lasagna made with zucchini instead of pasta",
+      prepTime: "40 minutes",
+      cookingTime: "35 minutes",
+      servings: 4,
+      cuisine: "Italian",
+      category: "Main Course",
+      ingredients: [
+        { quantity: "2", name: "zucchini", notes: "sliced lengthwise" },
+        { quantity: "1/2 lb", name: "ground beef" },
+        { quantity: "1/2 cup", name: "ricotta cheese" },
+        { quantity: "1/2 cup", name: "mozzarella cheese", notes: "shredded" },
+        { quantity: "1/2 cup", name: "tomato sauce" },
+        { quantity: "1/2 teaspoon", name: "oregano" },
+        { quantity: "1/2 teaspoon", name: "garlic powder" },
+      ],
+      instructions: [
+        "Preheat oven to 375°F (190°C).",
+        "Cook ground beef with tomato sauce and seasonings.",
+        "Layer zucchini slices, meat sauce, ricotta, and mozzarella.",
+        "Repeat layers and top with mozzarella.",
+        "Bake for 30–35 minutes until golden.",
+      ],
+      nutritionalInfo: {
+        calories: 310,
+        protein: "22g",
+        fat: "22g",
+        carbs: "8g",
+      },
+      dietaryInfo: ["Keto", "Low-Carb", "Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Beef",
+      cookingMethod: "Baking",
+      diet: "keto",
+    },
+    {
+      id: 79,
+      title: "Keto Chocolate Mousse",
+      image:
+        "https://healthyrecipesblogs.com/wp-content/uploads/2019/10/keto-chocolate-mousse-featured-2021.jpg",
+      rating: 4.9,
+      description: "Creamy chocolate mousse with no added sugar",
+      prepTime: "10 minutes",
+      cookingTime: "5 minutes (chill time extra)",
+      servings: 2,
+      cuisine: "French",
+      category: "Dessert",
+      ingredients: [
+        { quantity: "1/2 cup", name: "heavy cream" },
+        { quantity: "2 tablespoons", name: "cocoa powder" },
+        { quantity: "1 tablespoon", name: "erythritol" },
+        { quantity: "1/2 teaspoon", name: "vanilla extract" },
+        { quantity: "1/4 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "Whip heavy cream until soft peaks form.",
+        "Mix cocoa powder, erythritol, vanilla, and salt.",
+        "Fold dry mixture into whipped cream.",
+        "Chill for 30 minutes before serving.",
+      ],
+      nutritionalInfo: {
+        calories: 210,
+        protein: "3g",
+        fat: "20g",
+        carbs: "4g",
+      },
+      dietaryInfo: ["Keto", "Low-Carb", "Gluten-Free", "Vegetarian"],
+      mealType: ["Dessert"],
+      mainIngredient: "Milk",
+      cookingMethod: "No-Cook",
+      diet: "keto",
+    },
+  ],
+  vegetarian: [
+    {
+      id: 80,
+      title: "Vegetable Stir-Fry",
+      image:
+        "https://natashaskitchen.com/wp-content/uploads/2020/08/Vegetable-Stir-Fry-SQ.jpg",
+      rating: 4.8,
+      prepTime: "20 minutes",
+      cookingTime: "10 minutes",
+      servings: 2,
+      cuisine: "Asian",
+      category: "Main Course",
+      description: "Colorful stir-fried vegetables in a savory garlic sauce",
+      ingredients: [
+        { quantity: "1 cup", name: "broccoli", notes: "chopped" },
+        { quantity: "1 cup", name: "carrots", notes: "sliced" },
+        { quantity: "1 cup", name: "bell peppers", notes: "sliced" },
+        { quantity: "2 tablespoons", name: "soy sauce" },
+        { quantity: "1 tablespoon", name: "olive oil" },
+        { quantity: "2 cloves", name: "garlic", notes: "minced" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+      ],
+      instructions: [
+        "Heat olive oil in a pan over medium-high heat.",
+        "Add garlic and sauté for 1 minute until fragrant.",
+        "Add broccoli, carrots, and bell peppers, stir-fry for 5-6 minutes.",
+        "Pour in soy sauce and black pepper, cook for another 2-3 minutes.",
+        "Serve hot with rice or noodles.",
+      ],
+      nutritionalInfo: {
+        calories: 180,
+        protein: "6g",
+        fat: "8g",
+        carbs: "22g",
+        fiber: "5g",
+      },
+      dietaryInfo: ["Vegetarian", "Vegan", "Gluten-Free option"],
+      mealType: ["Lunch", "Dinner", "Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Stovetop",
+      diet: "vegetarian",
+    },
+    {
+      id: 81,
+      title: "Chickpea Salad",
+      image:
+        "https://www.themediterraneandish.com/wp-content/uploads/2023/12/TMD-Chickpea-Salad-Leads-01-Angle-Horizontal.jpg",
+      rating: 4.9,
+      prepTime: "10 minutes",
+      cookingTime: "0 minutes",
+      servings: 2,
+      cuisine: "Mediterranean",
+      category: "Salad",
+      description: "Refreshing and protein-packed chickpea salad",
+      ingredients: [
+        { quantity: "1 can", name: "chickpeas", notes: "drained and rinsed" },
+        { quantity: "1/2 cup", name: "cucumber", notes: "diced" },
+        { quantity: "1/2 cup", name: "tomatoes", notes: "diced" },
+        { quantity: "1/4 cup", name: "red onion", notes: "chopped" },
+        { quantity: "1 tablespoon", name: "olive oil" },
+        { quantity: "1 tablespoon", name: "lemon juice" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+      ],
+      instructions: [
+        "In a large bowl, combine chickpeas, cucumber, tomatoes, and red onion.",
+        "Drizzle olive oil and lemon juice over the salad.",
+        "Season with salt and toss to combine.",
+        "Serve chilled.",
+      ],
+      nutritionalInfo: {
+        calories: 250,
+        protein: "10g",
+        fat: "8g",
+        carbs: "36g",
+        fiber: "9g",
+      },
+      dietaryInfo: ["Vegetarian", "Vegan", "Gluten-Free"],
+      mealType: ["Lunch", "Snack"],
+      mainIngredient: "Dal",
+      cookingMethod: "No-Cook",
+      diet: "vegetarian",
+    },
+    {
+      id: 82,
+      title: "Lentil Soup",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTTwiwE3eGMsi3i-KSa2hzdolSWCEmSWGAg6w&s",
+      rating: 5.0,
+      prepTime: "35 minutes",
+      cookingTime: "25 minutes",
+      servings: 4,
+      cuisine: "Middle Eastern",
+      category: "Soup",
+      description: "Hearty and nutritious lentil soup with aromatic spices",
+      ingredients: [
+        { quantity: "1 cup", name: "red lentils", notes: "rinsed" },
+        { quantity: "1/2 cup", name: "carrots", notes: "diced" },
+        { quantity: "1/2 cup", name: "onion", notes: "chopped" },
+        { quantity: "2 cloves", name: "garlic", notes: "minced" },
+        { quantity: "1 teaspoon", name: "cumin powder" },
+        { quantity: "1/2 teaspoon", name: "turmeric" },
+        { quantity: "4 cups", name: "vegetable broth" },
+      ],
+      instructions: [
+        "Heat oil in a pot and sauté onion, garlic, and carrots until softened.",
+        "Add lentils, cumin, turmeric, and vegetable broth.",
+        "Bring to a boil, then reduce heat and simmer for 20-25 minutes.",
+        "Blend partially for a creamy texture if desired.",
+        "Serve hot with bread or rice.",
+      ],
+      nutritionalInfo: {
+        calories: 280,
+        protein: "16g",
+        fat: "4g",
+        carbs: "45g",
+        fiber: "12g",
+      },
+      dietaryInfo: ["Vegan", "Vegetarian", "Gluten-Free", "High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Dal",
+      cookingMethod: "Stovetop",
+      diet: "vegetarian",
+    },
+  ],
+  whole30: [
+    {
+      id: 83,
+      title: "Grilled Lemon Garlic Chicken",
+      image:
+        "https://www.tasteofhome.com/wp-content/uploads/2024/09/Easy-Grilled-Lemon-Chicken_EXPS_TOHD24_40183_JuliaHartbeck_6.jpg",
+      rating: 4.9,
+      prepTime: "25 minutes",
+      cookingTime: "20 minutes",
+      servings: 2,
+      cuisine: "American",
+      category: "Main Course",
+      description: "Juicy grilled chicken with lemon and garlic seasoning",
+      ingredients: [
+        { quantity: "2", name: "chicken breasts", notes: "boneless, skinless" },
+        { quantity: "2 tablespoons", name: "olive oil" },
+        { quantity: "2 cloves", name: "garlic", notes: "minced" },
+        { quantity: "1", name: "lemon", notes: "juiced" },
+        { quantity: "1 teaspoon", name: "paprika" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+      ],
+      instructions: [
+        "In a bowl, whisk together olive oil, garlic, lemon juice, paprika, salt, and pepper.",
+        "Coat chicken breasts in the marinade and let sit for at least 15 minutes.",
+        "Preheat grill to medium-high heat.",
+        "Grill chicken for 6-7 minutes per side, until cooked through.",
+        "Serve hot with lemon wedges.",
+      ],
+      nutritionalInfo: {
+        calories: 320,
+        protein: "36g",
+        fat: "18g",
+        carbs: "3g",
+      },
+      dietaryInfo: ["High-Protein", "Gluten-Free", "Low-Carb"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Grilling",
+      diet: "whole30",
+    },
+    {
+      id: 84,
+      title: "Sweet Potato Hash with Eggs",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhfgCjosgWJguqCC93JdL9RzsQ99RXRVbZWA&s",
+      rating: 5.0,
+      prepTime: "30 minutes",
+      cookingTime: "20 minutes",
+      servings: 2,
+      cuisine: "American",
+      category: "Breakfast",
+      description: "A nutritious sweet potato and egg breakfast hash",
+      ingredients: [
+        { quantity: "1", name: "sweet potato", notes: "diced" },
+        { quantity: "1/2 cup", name: "onion", notes: "chopped" },
+        { quantity: "1/2 cup", name: "bell peppers", notes: "chopped" },
+        { quantity: "1 tablespoon", name: "olive oil" },
+        { quantity: "2", name: "eggs" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+      ],
+      instructions: [
+        "Heat olive oil in a skillet over medium heat.",
+        "Add sweet potatoes and cook for 8-10 minutes until tender.",
+        "Add onions and bell peppers, cook for another 5 minutes.",
+        "Make small wells in the hash and crack eggs into them.",
+        "Cover and cook until eggs are set to desired doneness.",
+        "Season with salt and pepper, serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: 280,
+        protein: "10g",
+        fat: "12g",
+        carbs: "32g",
+        fiber: "6g",
+      },
+      dietaryInfo: ["Vegetarian", "Gluten-Free"],
+      mealType: ["Breakfast", "Brunch"],
+      mainIngredient: "Potato",
+      cookingMethod: "Stovetop",
+      diet: "whole30",
+    },
+    {
+      id: 85,
+      title: "Baked Salmon with Garlic and Herbs",
+      image:
+        "https://cdn.momsdish.com/wp-content/uploads/2021/05/Garlic-Herb-Roasted-Salmon011.jpg",
+      rating: 4.8,
+      prepTime: "25 minutes",
+      cookingTime: "20 minutes",
+      servings: 2,
+      cuisine: "Mediterranean",
+      category: "Main Course",
+      description: "Oven-baked salmon with a flavorful herb crust",
+      ingredients: [
+        { quantity: "2", name: "salmon fillets" },
+        { quantity: "1 tablespoon", name: "olive oil" },
+        { quantity: "2 cloves", name: "garlic", notes: "minced" },
+        { quantity: "1 teaspoon", name: "dried oregano" },
+        { quantity: "1/2 teaspoon", name: "salt" },
+        { quantity: "1/2 teaspoon", name: "black pepper" },
+        { quantity: "1", name: "lemon", notes: "sliced" },
+      ],
+      instructions: [
+        "Preheat oven to 375°F (190°C).",
+        "Place salmon fillets on a baking sheet lined with parchment paper.",
+        "Mix olive oil, garlic, oregano, salt, and pepper into a paste.",
+        "Spread mixture evenly over salmon fillets.",
+        "Top with lemon slices and bake for 18-20 minutes until salmon flakes easily.",
+        "Serve hot with a side salad or vegetables.",
+      ],
+      nutritionalInfo: {
+        calories: 350,
+        protein: "34g",
+        fat: "20g",
+        carbs: "4g",
+      },
+      dietaryInfo: ["High-Protein", "Gluten-Free", "Pescatarian", "Low-Carb"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Fish",
+      cookingMethod: "Baking",
+      diet: "whole30",
+    },
+  ],
+  japanese: [
+    {
+      id: 86,
+      title: "Sushi Rolls",
+      image:
+        "https://www.allrecipes.com/thmb/XT7-9MROYJZvNyQR4J40HGOVDmQ=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/19511smoked-salmon-sushi-rollfabeveryday4x3-159a22b4d3ac49fe9a146db94b53c930.jpg",
+      rating: 4.9,
+      prepTime: "40 minutes",
+      cookingTime: "10 minutes",
+      servings: 4,
+      cuisine: "Japanese",
+      category: "Main Course",
+      description:
+        "Traditional Japanese sushi rolls with rice, seaweed, and fresh fillings",
+      ingredients: [
+        { quantity: "2 cups", name: "sushi rice", notes: "cooked" },
+        { quantity: "4 sheets", name: "nori" },
+        { quantity: "1/2 cup", name: "rice vinegar" },
+        { quantity: "100g", name: "salmon", notes: "sliced" },
+        { quantity: "1/2", name: "cucumber", notes: "julienned" },
+      ],
+      instructions: [
+        "Season cooked sushi rice with rice vinegar.",
+        "Place a sheet of nori on a bamboo sushi mat.",
+        "Spread rice evenly on top, leaving a small border at the top edge.",
+        "Arrange salmon and cucumber in the center.",
+        "Roll tightly using the mat, then slice into pieces.",
+        "Serve with soy sauce, pickled ginger, and wasabi.",
+      ],
+      nutritionalInfo: {
+        calories: 280,
+        protein: "12g",
+        fat: "6g",
+        carbs: "46g",
+      },
+      dietaryInfo: ["Pescatarian", "Dairy-Free"],
+      mealType: ["Lunch", "Dinner", "Snack"],
+      mainIngredient: "Fish",
+      cookingMethod: "No-Cook",
+      diet: "whole30",
+    },
+    {
+      id: 87,
+      title: "Miso Soup",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTdYBr6d-_3otsb3-kN3Ujbi7peQriJx02SNA&s",
+      rating: 4.8,
+      prepTime: "15 minutes",
+      cookingTime: "10 minutes",
+      servings: 2,
+      cuisine: "Japanese",
+      category: "Soup",
+      description:
+        "A light and healthy soup made with miso paste, tofu, and seaweed",
+      ingredients: [
+        { quantity: "3 cups", name: "dashi broth" },
+        { quantity: "2 tablespoons", name: "miso paste" },
+        { quantity: "1/2 cup", name: "tofu", notes: "cubed" },
+        { quantity: "1 tablespoon", name: "wakame seaweed", notes: "dried" },
+        { quantity: "1", name: "green onion", notes: "chopped" },
+      ],
+      instructions: [
+        "Bring dashi broth to a gentle simmer.",
+        "Stir in miso paste until dissolved.",
+        "Add tofu cubes and wakame seaweed.",
+        "Simmer for 2-3 minutes until heated through.",
+        "Garnish with green onions and serve hot.",
+      ],
+      nutritionalInfo: {
+        calories: 90,
+        protein: "7g",
+        fat: "3g",
+        carbs: "10g",
+      },
+      dietaryInfo: ["Vegetarian", "Low-Calorie", "Gluten-Free (with GF miso)"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Tofu",
+      cookingMethod: "Stovetop",
+      diet: "whole30",
+    },
+    {
+      id: 88,
+      title: "Ramen",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQvJcmgBxotz7E_V0FcPu6tKHuW-SOEncooCQ&s",
+      rating: 4.9,
+      prepTime: "45 minutes",
+      cookingTime: "25 minutes",
+      servings: 2,
+      cuisine: "Japanese",
+      category: "Main Course",
+      description:
+        "Flavorful Japanese noodle soup with broth, toppings, and soft-boiled eggs",
+      ingredients: [
+        { quantity: "2 packs", name: "ramen noodles" },
+        { quantity: "3 cups", name: "chicken broth" },
+        { quantity: "2 tablespoons", name: "soy sauce" },
+        { quantity: "1", name: "soft-boiled egg", notes: "halved" },
+        { quantity: "1/2 cup", name: "sliced mushrooms" },
+      ],
+      instructions: [
+        "Bring chicken broth to a boil and add soy sauce.",
+        "Cook ramen noodles separately according to package instructions.",
+        "Add mushrooms to broth and simmer for 5 minutes.",
+        "Place cooked noodles in serving bowls and ladle broth over them.",
+        "Top with soft-boiled egg and additional toppings as desired.",
+      ],
+      nutritionalInfo: {
+        calories: 420,
+        protein: "18g",
+        fat: "12g",
+        carbs: "60g",
+      },
+      dietaryInfo: ["High-Carb", "Comfort Food"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Stovetop",
+      diet: "whole30",
+    },
+    {
+      id: 89,
+      title: "Teriyaki Chicken",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS3oDb5tGl5xDpuc1XTaHxk7lyNd42o8eoxIw&s",
+      rating: 4.7,
+      prepTime: "30 minutes",
+      cookingTime: "20 minutes",
+      servings: 2,
+      cuisine: "Japanese",
+      category: "Main Course",
+      description:
+        "Juicy chicken glazed with a sweet and savory teriyaki sauce",
+      ingredients: [
+        { quantity: "2", name: "chicken breasts", notes: "sliced" },
+        { quantity: "1/4 cup", name: "soy sauce" },
+        { quantity: "2 tablespoons", name: "mirin" },
+        { quantity: "2 tablespoons", name: "sugar" },
+        { quantity: "1 teaspoon", name: "ginger", notes: "grated" },
+      ],
+      instructions: [
+        "Mix soy sauce, mirin, sugar, and ginger in a bowl to make teriyaki sauce.",
+        "Heat a pan over medium heat and cook chicken until browned.",
+        "Pour sauce over chicken and simmer until thickened.",
+        "Serve with steamed rice or vegetables.",
+      ],
+      nutritionalInfo: {
+        calories: 350,
+        protein: "32g",
+        fat: "10g",
+        carbs: "30g",
+      },
+      dietaryInfo: ["High-Protein", "Dairy-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Stovetop",
+      diet: "whole30",
+    },
+    {
+      id: 90,
+      title: "Tempura",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR8ds7Qf2s131bbQp-4LgNaatBXtnIvzYnx6A&s",
+      rating: 4.8,
+      prepTime: "35 minutes",
+      cookingTime: "15 minutes",
+      servings: 3,
+      cuisine: "Japanese",
+      category: "Appetizer",
+      description: "Light and crispy battered vegetables and shrimp",
+      ingredients: [
+        { quantity: "1/2 cup", name: "flour" },
+        { quantity: "1/2 cup", name: "ice-cold water" },
+        { quantity: "1", name: "egg" },
+        { quantity: "100g", name: "shrimp", notes: "peeled" },
+        { quantity: "1/2 cup", name: "assorted vegetables", notes: "sliced" },
+      ],
+      instructions: [
+        "Prepare batter by whisking together flour, egg, and ice-cold water.",
+        "Dip shrimp and vegetables into the batter.",
+        "Fry in hot oil until golden and crispy.",
+        "Drain excess oil on paper towels and serve immediately with dipping sauce.",
+      ],
+      nutritionalInfo: {
+        calories: 300,
+        protein: "15g",
+        fat: "14g",
+        carbs: "28g",
+      },
+      dietaryInfo: ["Pescatarian"],
+      mealType: ["Appetizer", "Snack"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Frying",
+      diet: "whole30",
+    },
+    {
+      id: 91,
+      title: "Matcha Ice Cream",
+      image:
+        "https://salu-salo.com/wp-content/uploads/2019/06/Matcha-Ice-Cream-No-Churn-2.jpg",
+      rating: 5.0,
+      prepTime: "20 minutes + freezing time",
+      cookingTime: "5 minutes",
+      servings: 4,
+      cuisine: "Japanese",
+      category: "Dessert",
+      description: "Creamy and rich green tea-flavored ice cream",
+      ingredients: [
+        { quantity: "2 cups", name: "heavy cream" },
+        { quantity: "1 cup", name: "milk" },
+        { quantity: "1/2 cup", name: "sugar" },
+        { quantity: "2 tablespoons", name: "matcha powder" },
+        { quantity: "2", name: "egg yolks" },
+      ],
+      instructions: [
+        "Heat milk and sugar in a saucepan until sugar dissolves.",
+        "Whisk egg yolks in a bowl and slowly temper with warm milk mixture.",
+        "Add matcha powder and whisk until smooth.",
+        "Whip heavy cream until soft peaks form and fold into matcha mixture.",
+        "Freeze for at least 4 hours before serving.",
+      ],
+      nutritionalInfo: {
+        calories: 290,
+        protein: "5g",
+        fat: "18g",
+        carbs: "26g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Dessert"],
+      mainIngredient: "Milk",
+      cookingMethod: "No-Cook",
+      diet: "whole30",
+    },
+  ],
+  italian: [
+    {
+      id: 92,
+      title: "Margherita Pizza",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQqLEoGo3bHJYJEpRIaveeChxuCkXPoLk3RNg&s",
+      rating: 4.8,
+      prepTime: "30 minutes",
+      cookingTime: "15 minutes",
+      servings: 2,
+      cuisine: "Italian",
+      category: "Main Course",
+      description:
+        "Classic Italian pizza topped with fresh tomatoes, basil, and mozzarella",
+      ingredients: [
+        { quantity: "1", name: "pizza dough" },
+        { quantity: "1/2 cup", name: "tomato sauce" },
+        { quantity: "1 cup", name: "mozzarella", notes: "sliced" },
+        { quantity: "1 teaspoon", name: "olive oil" },
+        { quantity: "5-6", name: "fresh basil leaves" },
+      ],
+      instructions: [
+        "Preheat oven to 475°F (245°C).",
+        "Roll out pizza dough and place on a baking sheet or stone.",
+        "Spread tomato sauce evenly over dough.",
+        "Top with mozzarella slices and drizzle with olive oil.",
+        "Bake for 12–15 minutes until crust is golden and cheese is bubbly.",
+        "Garnish with fresh basil before serving.",
+      ],
+      nutritionalInfo: {
+        calories: 300,
+        protein: "12g",
+        fat: "10g",
+        carbs: "38g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Milk",
+      cookingMethod: "Baking",
+      diet: "whole30",
+    },
+    {
+      id: 93,
+      title: "Spaghetti Carbonara",
+      image:
+        "https://www.allrecipes.com/thmb/a_0W8yk_LLCtH-VPqg2uLD9I5Pk=/0x512/filters:no_upscale():max_bytes(150000):strip_icc()/11973-spaghetti-carbonara-ii-DDMFS-4x3-6edea51e421e4457ac0c3269f3be5157.jpg",
+      rating: 4.9,
+      prepTime: "25 minutes",
+      cookingTime: "15 minutes",
+      servings: 2,
+      cuisine: "Italian",
+      category: "Main Course",
+      description:
+        "Creamy pasta with pancetta, eggs, Parmesan, and black pepper",
+      ingredients: [
+        { quantity: "200g", name: "spaghetti" },
+        { quantity: "100g", name: "pancetta", notes: "diced" },
+        { quantity: "2", name: "eggs" },
+        { quantity: "1/2 cup", name: "Parmesan cheese", notes: "grated" },
+        { quantity: "1 teaspoon", name: "black pepper" },
+      ],
+      instructions: [
+        "Cook spaghetti in salted boiling water until al dente.",
+        "In a skillet, sauté pancetta until crisp.",
+        "Beat eggs with Parmesan cheese in a bowl.",
+        "Drain pasta and quickly toss with pancetta and pan drippings.",
+        "Remove from heat and stir in egg-cheese mixture to create a creamy sauce.",
+        "Season generously with black pepper and serve immediately.",
+      ],
+      nutritionalInfo: {
+        calories: 520,
+        protein: "22g",
+        fat: "20g",
+        carbs: "62g",
+      },
+      dietaryInfo: ["High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Pasta",
+      cookingMethod: "Stovetop",
+      diet: "whole30",
+    },
+    {
+      id: 94,
+      title: "Minestrone Soup",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSlLnTDF5PsCsyCfxn6tDxJYdDqwkYX9OMc_A&s",
+      rating: 4.6,
+      prepTime: "40 minutes",
+      cookingTime: "30 minutes",
+      servings: 4,
+      cuisine: "Italian",
+      category: "Soup",
+      description: "Hearty vegetable soup with beans, pasta, and Italian herbs",
+      ingredients: [
+        { quantity: "1", name: "onion", notes: "chopped" },
+        { quantity: "2", name: "carrots", notes: "diced" },
+        { quantity: "1 cup", name: "kidney beans" },
+        { quantity: "1/2 cup", name: "small pasta" },
+        { quantity: "1 teaspoon", name: "Italian seasoning" },
+      ],
+      instructions: [
+        "Heat oil in a large pot and sauté onions and carrots until softened.",
+        "Add kidney beans, pasta, and Italian seasoning.",
+        "Pour in water or broth and bring to a boil.",
+        "Simmer until vegetables and pasta are cooked through.",
+        "Adjust seasoning and serve warm with bread.",
+      ],
+      nutritionalInfo: {
+        calories: 250,
+        protein: "9g",
+        fat: "5g",
+        carbs: "45g",
+      },
+      dietaryInfo: ["Vegetarian", "High-Fiber"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Vegetable",
+      cookingMethod: "Stovetop",
+      diet: "whole30",
+    },
+    {
+      id: 95,
+      title: "Risotto alla Milanese",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRQ8HC6HMLhBnUUcBbE1smE5tKpo0BMtUdz4w&s",
+      rating: 4.8,
+      prepTime: "35 minutes",
+      cookingTime: "25 minutes",
+      servings: 3,
+      cuisine: "Italian",
+      category: "Main Course",
+      description: "Creamy saffron-infused risotto with Parmesan and butter",
+      ingredients: [
+        { quantity: "1 cup", name: "Arborio rice" },
+        { quantity: "4 cups", name: "chicken broth" },
+        { quantity: "1/2 teaspoon", name: "saffron" },
+        { quantity: "1/2 cup", name: "Parmesan cheese", notes: "grated" },
+        { quantity: "2 tablespoons", name: "butter" },
+      ],
+      instructions: [
+        "Heat broth in a saucepan and keep warm.",
+        "In a large pan, sauté Arborio rice briefly in butter.",
+        "Add saffron and ladle in broth gradually, stirring constantly.",
+        "Continue adding broth until rice is creamy and al dente.",
+        "Stir in Parmesan cheese before serving.",
+      ],
+      nutritionalInfo: {
+        calories: 410,
+        protein: "12g",
+        fat: "15g",
+        carbs: "56g",
+      },
+      dietaryInfo: ["Vegetarian (if vegetable broth used)"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Rice",
+      cookingMethod: "Stovetop",
+      diet: "whole30",
+    },
+    {
+      id: 96,
+      title: "Tiramisu",
+      image:
+        "https://i0.wp.com/mediterraneantaste.com/wp-content/uploads/2023/11/tiramisu-4583.jpg?fit=1024%2C683&ssl=1",
+      rating: 5.0,
+      prepTime: "20 minutes + chilling time",
+      cookingTime: "0 minutes",
+      servings: 6,
+      cuisine: "Italian",
+      category: "Dessert",
+      description:
+        "Classic Italian dessert made with layers of coffee-soaked ladyfingers and mascarpone cream",
+      ingredients: [
+        { quantity: "200g", name: "ladyfingers" },
+        { quantity: "1 cup", name: "espresso" },
+        { quantity: "250g", name: "mascarpone cheese" },
+        { quantity: "2", name: "eggs" },
+        { quantity: "2 tablespoons", name: "cocoa powder" },
+      ],
+      instructions: [
+        "Dip ladyfingers quickly into espresso and layer them in a dish.",
+        "Whisk mascarpone with egg yolks until smooth.",
+        "Whip egg whites until stiff peaks form, then fold into mascarpone mixture.",
+        "Spread cream over ladyfingers and repeat layers.",
+        "Dust with cocoa powder and refrigerate for at least 4 hours before serving.",
+      ],
+      nutritionalInfo: {
+        calories: 450,
+        protein: "8g",
+        fat: "28g",
+        carbs: "42g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Dessert"],
+      mainIngredient: "Milk",
+      cookingMethod: "No-Cook",
+      diet: "whole30",
+    },
+  ],
+  mexican: [
+    {
+      id: 97,
+      title: "Tacos al Pastor",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTXiNsnnAfPLpQdWnxXVBCSpen0f_Prp31lig&s",
+      rating: 4.9,
+      prepTime: "30 minutes",
+      cookingTime: "20 minutes",
+      servings: 4,
+      cuisine: "Mexican",
+      category: "Main Course",
+      description:
+        "Traditional Mexican tacos with marinated pork, pineapple, and onions",
+      ingredients: [
+        { quantity: "500g", name: "pork", notes: "thinly sliced" },
+        { quantity: "1/2 cup", name: "pineapple", notes: "diced" },
+        { quantity: "1 teaspoon", name: "cumin" },
+        { quantity: "1 teaspoon", name: "paprika" },
+        { quantity: "8", name: "corn tortillas" },
+      ],
+      instructions: [
+        "Marinate sliced pork with cumin, paprika, and pineapple for at least 30 minutes.",
+        "Cook pork in a hot skillet until browned and tender.",
+        "Warm corn tortillas on a griddle.",
+        "Fill tortillas with pork and pineapple.",
+        "Top with onions, cilantro, and a squeeze of lime before serving.",
+      ],
+      nutritionalInfo: {
+        calories: 420,
+        protein: "28g",
+        fat: "18g",
+        carbs: "38g",
+      },
+      dietaryInfo: ["High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Pork",
+      cookingMethod: "Stovetop",
+    },
+    {
+      id: 98,
+      title: "Guacamole",
+      image:
+        "https://www.allrecipes.com/thmb/6RyFPH5N4KKmZhNY0Giob_Jj3wc=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/AR-14231-guacamole-4x3-f7a3b5752c7f4f3fb934d03a8b548826.jpg",
+      rating: 4.8,
+      prepTime: "10 minutes",
+      cookingTime: "0 minutes",
+      servings: 4,
+      cuisine: "Mexican",
+      category: "Appetizer",
+      description: "Creamy avocado dip with lime, tomatoes, and cilantro",
+      ingredients: [
+        { quantity: "2", name: "avocados", notes: "mashed" },
+        { quantity: "1", name: "lime", notes: "juiced" },
+        { quantity: "1", name: "tomato", notes: "diced" },
+        { quantity: "1/4 cup", name: "onion", notes: "chopped" },
+        { quantity: "1 tablespoon", name: "cilantro", notes: "chopped" },
+      ],
+      instructions: [
+        "Mash avocados in a bowl.",
+        "Stir in lime juice, tomato, onion, and cilantro.",
+        "Mix until smooth but chunky.",
+        "Season with salt to taste and serve with tortilla chips.",
+      ],
+      nutritionalInfo: {
+        calories: 150,
+        protein: "2g",
+        fat: "13g",
+        carbs: "8g",
+      },
+      dietaryInfo: ["Vegan", "Gluten-Free"],
+      mealType: ["Appetizer", "Snack"],
+      mainIngredient: "Fruit",
+      cookingMethod: "No-Cook",
+    },
+    {
+      id: 99,
+      title: "Chiles Rellenos",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcT2NdusYL2SpEF52z1TEihwWUM6A9JHmAoNPg&s",
+      rating: 4.7,
+      prepTime: "40 minutes",
+      cookingTime: "25 minutes",
+      servings: 4,
+      cuisine: "Mexican",
+      category: "Main Course",
+      description:
+        "Poblano peppers stuffed with cheese and coated in a light egg batter",
+      ingredients: [
+        { quantity: "4", name: "poblano peppers" },
+        { quantity: "1 cup", name: "cheese", notes: "shredded" },
+        { quantity: "2", name: "eggs", notes: "separated" },
+        { quantity: "1/2 cup", name: "flour" },
+        { quantity: "1/2 cup", name: "tomato sauce" },
+      ],
+      instructions: [
+        "Roast poblano peppers until charred, then peel and remove seeds.",
+        "Stuff peppers with shredded cheese.",
+        "Beat egg whites until fluffy, then fold in yolks to make batter.",
+        "Coat peppers in flour, then dip in egg batter.",
+        "Fry until golden and serve with tomato sauce.",
+      ],
+      nutritionalInfo: {
+        calories: 380,
+        protein: "15g",
+        fat: "22g",
+        carbs: "28g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Cheese",
+      cookingMethod: "Frying",
+    },
+    {
+      id: 100,
+      title: "Enchiladas Verdes",
+      image:
+        "https://hips.hearstapps.com/hmg-prod/images/enchiladas-verdes-recipe-2-1659537049.jpg?crop=0.6666666666666667xw:1xh;center,top&resize=1200:*",
+      rating: 4.9,
+      prepTime: "35 minutes",
+      cookingTime: "25 minutes",
+      servings: 4,
+      cuisine: "Mexican",
+      category: "Main Course",
+      description:
+        "Corn tortillas filled with chicken and topped with green tomatillo sauce",
+      ingredients: [
+        { quantity: "8", name: "corn tortillas" },
+        { quantity: "2 cups", name: "chicken", notes: "shredded" },
+        { quantity: "1 cup", name: "tomatillo sauce" },
+        { quantity: "1/2 cup", name: "sour cream" },
+        { quantity: "1/2 cup", name: "cheese", notes: "grated" },
+      ],
+      instructions: [
+        "Warm tortillas and fill with shredded chicken.",
+        "Roll and place in a baking dish.",
+        "Pour tomatillo sauce over tortillas and sprinkle with cheese.",
+        "Bake at 375°F (190°C) for 20 minutes.",
+        "Top with sour cream before serving.",
+      ],
+      nutritionalInfo: {
+        calories: 450,
+        protein: "28g",
+        fat: "20g",
+        carbs: "40g",
+      },
+      dietaryInfo: ["High-Protein"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Chicken",
+      cookingMethod: "Baking",
+    },
+    {
+      id: 101,
+      title: "Tamales",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTwjPNq42hUDYUlC8C-MpbVMjaAG9UO7PZ-1Q&s",
+      rating: 4.8,
+      prepTime: "1 hour",
+      cookingTime: "1 hour",
+      servings: 6,
+      cuisine: "Mexican",
+      category: "Main Course",
+      description:
+        "Corn masa filled with meat or cheese, wrapped in corn husks and steamed",
+      ingredients: [
+        { quantity: "2 cups", name: "masa harina" },
+        { quantity: "1 cup", name: "chicken broth" },
+        { quantity: "1/2 cup", name: "lard" },
+        { quantity: "1 cup", name: "chicken", notes: "shredded" },
+        { quantity: "10", name: "corn husks", notes: "soaked" },
+      ],
+      instructions: [
+        "Soak corn husks in warm water until pliable.",
+        "Mix masa harina with broth and lard to form dough.",
+        "Spread dough on corn husks, fill with chicken, and fold.",
+        "Steam tamales for 1 hour until firm.",
+        "Serve warm with salsa.",
+      ],
+      nutritionalInfo: {
+        calories: 320,
+        protein: "12g",
+        fat: "15g",
+        carbs: "34g",
+      },
+      dietaryInfo: ["Gluten-Free"],
+      mealType: ["Lunch", "Dinner"],
+      mainIngredient: "Corn",
+      cookingMethod: "Steaming",
+    },
+    {
+      id: 102,
+      title: "Churros",
+      image:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQZpJgGjJcAB2M9aWSUPMFFC4ZGsYZrS7GKqw&s",
+      rating: 5.0,
+      prepTime: "25 minutes",
+      cookingTime: "15 minutes",
+      servings: 6,
+      cuisine: "Mexican",
+      category: "Dessert",
+      description:
+        "Crispy fried dough coated in cinnamon sugar, served with chocolate sauce",
+      ingredients: [
+        { quantity: "1 cup", name: "flour" },
+        { quantity: "1 cup", name: "water" },
+        { quantity: "2 tablespoons", name: "butter" },
+        { quantity: "1/2 cup", name: "sugar" },
+        { quantity: "1 teaspoon", name: "cinnamon" },
+      ],
+      instructions: [
+        "Bring water and butter to a boil in a saucepan.",
+        "Stir in flour to form dough.",
+        "Pipe dough into hot oil and fry until golden.",
+        "Roll churros in cinnamon sugar.",
+        "Serve warm with chocolate sauce for dipping.",
+      ],
+      nutritionalInfo: {
+        calories: 280,
+        protein: "4g",
+        fat: "12g",
+        carbs: "38g",
+      },
+      dietaryInfo: ["Vegetarian"],
+      mealType: ["Dessert", "Snack"],
+      mainIngredient: "Flour",
+      cookingMethod: "Frying",
+    },
+  ],
+};
+
+export default recipes;
