+++ frontend/src/pages/Home.jsx
@@ -1,5 +1,4 @@
 import "../App.css";
-import "./recipes";
 import Carousel from "../components/home_page_components/Carousel";
 import Testimonial from "../components/home_page_components/Testimonial";
 import RecipeSection from "../components/home_page_components/RecipeSection";
