+++ frontend/src/pages/categories/Appetizer.jsx
@@ -1,112 +1,9 @@
 import React, { useState } from 'react';
 import '../../styles/AllCat.css';
+import recipes from '../../data/recipes';
 
 const Appetizer = () => {
     const [setSelectedRecipe] = useState(null);
-
-    const appetizerRecipes = [
-        {
-          id: 1,
-          title: "Samosa",
-          image: "https://static.toiimg.com/thumb/61050397.cms?imgsize=246859&width=800&height=800",
-          rating: 4.9,
-          prepTime: "45 minutes",
-          description: "Crispy, deep-fried pastry filled with spiced potatoes and peas",
-          ingredients: [
-            { quantity: "2 cups", name: "all-purpose flour" },
-            { quantity: "4 tablespoons", name: "oil" },
-            { quantity: "1/2 teaspoon", name: "salt" },
-            { quantity: "3", name: "potatoes", notes: "boiled and mashed" },
-            { quantity: "1/2 cup", name: "peas" },
-            { quantity: "1 teaspoon", name: "cumin seeds" },
-            { quantity: "1 teaspoon", name: "garam masala" }
-          ]
-        },
-        {
-          id: 2,
-          title: "Paneer Tikka",
-          image: "https://orders.popskitchen.in/storage/2024/09/image-358.png",
-          rating: 4.8,
-          prepTime: "30 minutes",
-          description: "Grilled spiced paneer cubes with bell peppers and onions",
-          ingredients: [
-            { quantity: "250 grams", name: "paneer", notes: "cubed" },
-            { quantity: "1/2 cup", name: "yogurt" },
-            { quantity: "1 teaspoon", name: "red chili powder" },
-            { quantity: "1 teaspoon", name: "garam masala" },
-            { quantity: "1 teaspoon", name: "lemon juice" },
-            { quantity: "1 tablespoon", name: "ginger-garlic paste" },
-            { quantity: "1", name: "bell pepper", notes: "cubed" }
-          ]
-        },
-        {
-          id: 3,
-          title: "Vegetable Pakora",
-          image: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSqxHWAXstc2X1NLPKNstpaRJs-uU4Rb_ESmQ&s",
-          rating: 4.7,
-          prepTime: "20 minutes",
-          description: "Crispy deep-fried vegetable fritters made with gram flour",
-          ingredients: [
-            { quantity: "1 cup", name: "gram flour (besan)" },
-            { quantity: "1", name: "potato", notes: "thinly sliced" },
-            { quantity: "1", name: "onion", notes: "sliced" },
-            { quantity: "1/2 teaspoon", name: "turmeric powder" },
-            { quantity: "1 teaspoon", name: "red chili powder" },
-            { quantity: "1 teaspoon", name: "cumin seeds" },
-            { quantity: "1/2 cup", name: "water" }
-          ]
-        },
-        {
-          id: 4,
-          title: "Dahi Puri",
-          image: "https://hogr.app/blog/wp-content/uploads/2022/10/dahi-puri-chat-is-indian-road-side-snack-item-which-is-especially-popular-state-maharashtra-india-scaled.webp",
-          rating: 4.9,
-          prepTime: "25 minutes",
-          description: "Crispy puris filled with yogurt, chutneys, and spiced potatoes",
-          ingredients: [
-            { quantity: "10", name: "pani puris" },
-            { quantity: "1 cup", name: "yogurt", notes: "whisked" },
-            { quantity: "1", name: "potato", notes: "boiled and mashed" },
-            { quantity: "2 tablespoons", name: "tamarind chutney" },
-            { quantity: "2 tablespoons", name: "mint chutney" },
-            { quantity: "1/2 teaspoon", name: "chaat masala" },
-            { quantity: "1 tablespoon", name: "sev" }
-          ]
-        },
-        {
-          id: 5,
-          title: "Hara Bhara Kabab",
-          image: "https://www.vegrecipesofindia.com/wp-content/uploads/2013/01/hara-bhara-kabab-recipe-1-500x500.jpg",
-          rating: 4.8,
-          prepTime: "30 minutes",
-          description: "Healthy pan-fried spinach and pea patties with spices",
-          ingredients: [
-            { quantity: "1 cup", name: "spinach", notes: "blanched" },
-            { quantity: "1/2 cup", name: "peas", notes: "boiled" },
-            { quantity: "1", name: "potato", notes: "boiled and mashed" },
-            { quantity: "1 tablespoon", name: "ginger-garlic paste" },
-            { quantity: "1 teaspoon", name: "coriander powder" },
-            { quantity: "1 teaspoon", name: "green chili", notes: "chopped" },
-            { quantity: "2 tablespoons", name: "gram flour" }
-          ]
-        },
-        {
-          id: 6,
-          title: "Masala Papad",
-          image: "https://www.whiskaffair.com/wp-content/uploads/2020/09/Masala-Papad-2-3.jpg",
-          rating: 4.7,
-          prepTime: "10 minutes",
-          description: "Crispy roasted papad topped with spicy onions and tomatoes",
-          ingredients: [
-            { quantity: "2", name: "papads", notes: "roasted" },
-            { quantity: "1", name: "onion", notes: "finely chopped" },
-            { quantity: "1", name: "tomato", notes: "finely chopped" },
-            { quantity: "1/2 teaspoon", name: "red chili powder" },
-            { quantity: "1 teaspoon", name: "lemon juice" },
-            { quantity: "1 tablespoon", name: "coriander leaves", notes: "chopped" }
-          ]
-        }
-      ];
     
       return (
         <div className="desserts-wrapper">
@@ -132,7 +29,7 @@ const Appetizer = () => {
           </div>
     
           <div className="recipes-showcase">
-            {appetizerRecipes.map(recipe => (
+            {recipes["appetizers"].map((recipe) => (
               <div key={recipe.id} className="recipe-card">
                 <div className="card-media">
                   <img src={recipe.image} alt={recipe.title} />
