+++ frontend/src/pages/categories/Japanese.jsx
@@ -1,102 +1,10 @@
 import React, { useState } from 'react';
 import '../../styles/AllCat.css';
+import recipes from '../../data/recipes';
 
 const Japanese = () => {
     const [setSelectedRecipe] = useState(null);
 
-    const japaneseRecipes = [
-        {
-          id: 1,
-          title: "Sushi Rolls",
-          image: "https://www.allrecipes.com/thmb/XT7-9MROYJZvNyQR4J40HGOVDmQ=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/19511smoked-salmon-sushi-rollfabeveryday4x3-159a22b4d3ac49fe9a146db94b53c930.jpg",
-          rating: 4.9,
-          prepTime: "40 minutes",
-          description: "Traditional Japanese sushi rolls with rice, seaweed, and fresh fillings",
-          ingredients: [
-            { quantity: "2 cups", name: "sushi rice", notes: "cooked" },
-            { quantity: "4 sheets", name: "nori" },
-            { quantity: "1/2 cup", name: "rice vinegar" },
-            { quantity: "100g", name: "salmon", notes: "sliced" },
-            { quantity: "1/2", name: "cucumber", notes: "julienned" }
-          ]
-        },
-        {
-          id: 2,
-          title: "Miso Soup",
-          image: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTdYBr6d-_3otsb3-kN3Ujbi7peQriJx02SNA&s",
-          rating: 4.8,
-          prepTime: "15 minutes",
-          description: "A light and healthy soup made with miso paste, tofu, and seaweed",
-          ingredients: [
-            { quantity: "3 cups", name: "dashi broth" },
-            { quantity: "2 tablespoons", name: "miso paste" },
-            { quantity: "1/2 cup", name: "tofu", notes: "cubed" },
-            { quantity: "1 tablespoon", name: "wakame seaweed", notes: "dried" },
-            { quantity: "1", name: "green onion", notes: "chopped" }
-          ]
-        },
-        {
-          id: 3,
-          title: "Ramen",
-          image: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQvJcmgBxotz7E_V0FcPu6tKHuW-SOEncooCQ&s",
-          rating: 4.9,
-          prepTime: "45 minutes",
-          description: "Flavorful Japanese noodle soup with broth, toppings, and soft-boiled eggs",
-          ingredients: [
-            { quantity: "2 packs", name: "ramen noodles" },
-            { quantity: "3 cups", name: "chicken broth" },
-            { quantity: "2 tablespoons", name: "soy sauce" },
-            { quantity: "1", name: "soft-boiled egg", notes: "halved" },
-            { quantity: "1/2 cup", name: "sliced mushrooms" }
-          ]
-        },
-        {
-          id: 4,
-          title: "Teriyaki Chicken",
-          image: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS3oDb5tGl5xDpuc1XTaHxk7lyNd42o8eoxIw&s",
-          rating: 4.7,
-          prepTime: "30 minutes",
-          description: "Juicy chicken glazed with a sweet and savory teriyaki sauce",
-          ingredients: [
-            { quantity: "2", name: "chicken breasts", notes: "sliced" },
-            { quantity: "1/4 cup", name: "soy sauce" },
-            { quantity: "2 tablespoons", name: "mirin" },
-            { quantity: "2 tablespoons", name: "sugar" },
-            { quantity: "1 teaspoon", name: "ginger", notes: "grated" }
-          ]
-        },
-        {
-          id: 5,
-          title: "Tempura",
-          image: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR8ds7Qf2s131bbQp-4LgNaatBXtnIvzYnx6A&s",
-          rating: 4.8,
-          prepTime: "35 minutes",
-          description: "Light and crispy battered vegetables and shrimp",
-          ingredients: [
-            { quantity: "1/2 cup", name: "flour" },
-            { quantity: "1/2 cup", name: "ice-cold water" },
-            { quantity: "1", name: "egg" },
-            { quantity: "100g", name: "shrimp", notes: "peeled" },
-            { quantity: "1/2 cup", name: "assorted vegetables", notes: "sliced" }
-          ]
-        },
-        {
-          id: 6,
-          title: "Matcha Ice Cream",
-          image: "https://salu-salo.com/wp-content/uploads/2019/06/Matcha-Ice-Cream-No-Churn-2.jpg",
-          rating: 5.0,
-          prepTime: "20 minutes + freezing time",
-          description: "Creamy and rich green tea-flavored ice cream",
-          ingredients: [
-            { quantity: "2 cups", name: "heavy cream" },
-            { quantity: "1 cup", name: "milk" },
-            { quantity: "1/2 cup", name: "sugar" },
-            { quantity: "2 tablespoons", name: "matcha powder" },
-            { quantity: "2", name: "egg yolks" }
-          ]
-        }
-      ];
-    
       return (
         <div className="desserts-wrapper">
           <div className="hero-section">
@@ -121,7 +29,7 @@ const Japanese = () => {
           </div>
     
           <div className="recipes-showcase">
-            {japaneseRecipes.map(recipe => (
+            {recipes["japanese"].map(recipe => (
               <div key={recipe.id} className="recipe-card">
                 <div className="card-media">
                   <img src={recipe.image} alt={recipe.title} />
