+++ frontend/src/pages/categories/whole30.jsx
@@ -1,113 +1,9 @@
 import React, { useState } from 'react';
 import '../../styles/AllCat.css';
+import recipes from '../../data/recipes';
 
 const Whole30 = () => {
     const [setSelectedRecipe] = useState(null);
-
-    const whole30Recipes = [
-        {
-          id: 1,
-          title: "Grilled Lemon Garlic Chicken",
-          image: "https://www.tasteofhome.com/wp-content/uploads/2024/09/Easy-Grilled-Lemon-Chicken_EXPS_TOHD24_40183_JuliaHartbeck_6.jpg",
-          rating: 4.9,
-          prepTime: "25 minutes",
-          description: "Juicy grilled chicken with lemon and garlic seasoning",
-          ingredients: [
-            { quantity: "2", name: "chicken breasts", notes: "boneless, skinless" },
-            { quantity: "2 tablespoons", name: "olive oil" },
-            { quantity: "2 cloves", name: "garlic", notes: "minced" },
-            { quantity: "1", name: "lemon", notes: "juiced" },
-            { quantity: "1 teaspoon", name: "paprika" },
-            { quantity: "1/2 teaspoon", name: "salt" },
-            { quantity: "1/2 teaspoon", name: "black pepper" }
-          ]
-        },
-        {
-          id: 2,
-          title: "Cauliflower Rice Stir-Fry",
-          image: "https://cdn.loveandlemons.com/wp-content/uploads/2023/01/cauliflower-fried-rice.jpg",
-          rating: 4.8,
-          prepTime: "20 minutes",
-          description: "A delicious, low-carb stir-fry with cauliflower rice",
-          ingredients: [
-            { quantity: "2 cups", name: "cauliflower rice" },
-            { quantity: "1/2 cup", name: "carrots", notes: "diced" },
-            { quantity: "1/2 cup", name: "bell peppers", notes: "sliced" },
-            { quantity: "1/4 cup", name: "onion", notes: "chopped" },
-            { quantity: "1 tablespoon", name: "coconut aminos" },
-            { quantity: "1 tablespoon", name: "olive oil" },
-            { quantity: "1 teaspoon", name: "ginger", notes: "grated" }
-          ]
-        },
-        {
-          id: 3,
-          title: "Zucchini Noodles with Avocado Pesto",
-          image: "https://pinchofyum.com/wp-content/uploads/Zucchini-Spaghetti-6.jpg",
-          rating: 4.9,
-          prepTime: "15 minutes",
-          description: "Healthy zucchini noodles tossed in creamy avocado pesto",
-          ingredients: [
-            { quantity: "2", name: "zucchinis", notes: "spiralized" },
-            { quantity: "1", name: "avocado" },
-            { quantity: "1/4 cup", name: "fresh basil" },
-            { quantity: "2 tablespoons", name: "olive oil" },
-            { quantity: "1", name: "garlic clove", notes: "minced" },
-            { quantity: "1/2 teaspoon", name: "salt" },
-            { quantity: "1/2 teaspoon", name: "lemon juice" }
-          ]
-        },
-        {
-          id: 4,
-          title: "Sweet Potato Hash with Eggs",
-          image: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhfgCjosgWJguqCC93JdL9RzsQ99RXRVbZWA&s",
-          rating: 5.0,
-          prepTime: "30 minutes",
-          description: "A nutritious sweet potato and egg breakfast hash",
-          ingredients: [
-            { quantity: "1", name: "sweet potato", notes: "diced" },
-            { quantity: "1/2 cup", name: "onion", notes: "chopped" },
-            { quantity: "1/2 cup", name: "bell peppers", notes: "chopped" },
-            { quantity: "1 tablespoon", name: "olive oil" },
-            { quantity: "2", name: "eggs" },
-            { quantity: "1/2 teaspoon", name: "salt" },
-            { quantity: "1/2 teaspoon", name: "black pepper" }
-          ]
-        },
-        {
-          id: 5,
-          title: "Baked Salmon with Garlic and Herbs",
-          image: "https://cdn.momsdish.com/wp-content/uploads/2021/05/Garlic-Herb-Roasted-Salmon011.jpg",
-          rating: 4.8,
-          prepTime: "25 minutes",
-          description: "Oven-baked salmon with a flavorful herb crust",
-          ingredients: [
-            { quantity: "2", name: "salmon fillets" },
-            { quantity: "1 tablespoon", name: "olive oil" },
-            { quantity: "2 cloves", name: "garlic", notes: "minced" },
-            { quantity: "1 teaspoon", name: "dried oregano" },
-            { quantity: "1/2 teaspoon", name: "salt" },
-            { quantity: "1/2 teaspoon", name: "black pepper" },
-            { quantity: "1", name: "lemon", notes: "sliced" }
-          ]
-        },
-        {
-          id: 6,
-          title: "Coconut Curry Chicken",
-          image: "https://www.wellplated.com/wp-content/uploads/2019/07/Slow-Cooker-Chicken-Curry-Recipe.jpg",
-          rating: 4.9,
-          prepTime: "35 minutes",
-          description: "Rich and creamy coconut curry chicken",
-          ingredients: [
-            { quantity: "2", name: "chicken breasts", notes: "cubed" },
-            { quantity: "1 cup", name: "coconut milk" },
-            { quantity: "1/2 cup", name: "tomato puree" },
-            { quantity: "1 tablespoon", name: "curry powder" },
-            { quantity: "1/2 teaspoon", name: "turmeric" },
-            { quantity: "1/2 teaspoon", name: "salt" },
-            { quantity: "1 tablespoon", name: "olive oil" }
-          ]
-        }
-      ];
     
       return (
         <div className="desserts-wrapper">
@@ -133,7 +29,7 @@ const Whole30 = () => {
           </div>
     
           <div className="recipes-showcase">
-            {whole30Recipes.map(recipe => (
+            {recipes["whole30"].map(recipe => (
               <div key={recipe.id} className="recipe-card">
                 <div className="card-media">
                   <img src={recipe.image} alt={recipe.title} />
